Core Functions

Quiqbox.getCharge — Method

getCharge(nucs::Union{Vector{String}, Tuple{Vararg{String}}}) -> Int

Return the total electric charge (in 𝑒) of the input nuclei.

Quiqbox.orbitalLin — Function

orbitalLin(subshell::String, D::Int=3) -> Tuple{Vararg{NTuple{3, Int}}}

Return all the possible angular momentum configuration(s) within the input subshell of D dimension.

Quiqbox.changeMapping — Method

changeMapping(pb::ParamBox, mapFunction::Function, outputName::Symbol=V; 
              canDiff::Bool=true) -> 
ParamBox{T, outputName}

Change the mapping function of pb. The name of the output variable of the returned ParamBox can be specified by outputName, and its differentiability is determined by canDiff.

Quiqbox.dataOf — Method

dataOf(pb::ParamBox{T}) where {T} -> Array{T, 0}

Return the 0-D Array containing data stored in pb.

Quiqbox.disableDiff! — Method

disableDiff!(pb::ParamBox) -> ParamBox

Mark the pb as "non-differentiable" and then return it.

Quiqbox.enableDiff! — Method

enableDiff!(pb::ParamBox) -> ParamBox

Mark the input pb as "differentiable" and then return it.

Quiqbox.getVar — Function

getVar(pb::ParamBox{T}, forDifferentiation::Bool=false) -> Symbol

Return the name of (the output variable of) pb. If forDifferentiation is set to true, the name of the independent variable held by pb is returned.

Quiqbox.getVarDict — Method

getVarDict(obj::Union{ParamBox, Tuple{Vararg{ParamBox}}, AbstractArray{<:ParamBox}}) -> 
Dict{Symbol}

Return a Dict that stores the independent variable(s) of the parameter container(s) and its(their) corresponding value(s).

NOTE: Once obj is mutated, the generated Dict may no longer be up to date.

Quiqbox.inSymOf — Method

inSymOf(pb::ParamBox) -> Symbol

Return the name (with the index if available) of the input variable of pb.

Quiqbox.inSymOfCore — Method

inSymOfCore(pb::ParamBox) -> Symbol

Return the Symbol of the input variable of pb.

Quiqbox.inSymValOf — Method

inSymValOf(pb::ParamBox{T}) where {T} -> ::Pair{Symbol, T}

Return a Pair of the name (with the index if available) and the value of the input variable of pb.

Quiqbox.inValOf — Method

inValOf(pb::ParamBox{T}) where {T} -> T

Return the value of the input variable of pb. Equivalent to pb[].

Quiqbox.inVarCopy — Method

inVarCopy(pb::ParamBox) -> ParamBox{<:Number, <:Any, FLevel{0}}

Return a new ParamBox of which the input variable is a shallow copy of the input variable of pb.

≡≡≡ Example(s) ≡≡≡

julia> pb1 = ParamBox(-2.0, :a, abs)
ParamBox{Float64, :a, FLevel{1}}(-2.0)[∂][x_a]

julia> pb2 = inVarCopy(pb1)
ParamBox{Float64, :x_a, FLevel{0}}(-2.0)[∂][x_a]

julia> pb1[] == pb2[] == -2.0
true

julia> pb1[] = 1.1
1.1

julia> pb2[]
1.1

Quiqbox.isDiffParam — Method

isDiffParam(pb::ParamBox) -> Bool

Return the Boolean value of if pb is differentiable.

Quiqbox.mapOf — Method

mapOf(pb::ParamBox) -> Function

Return the mapping function of pb.

Quiqbox.outSymOf — Method

outSymOf(pb::ParamBox) -> Symbol

Return the name (with the index if available) of the output variable of pb.

Quiqbox.outSymOfCore — Method

outSymOfCore(pb::ParamBox) -> Symbol

Return the Symbol of the output variable of pb.

Quiqbox.outSymValOf — Method

outSymValOf(pb::ParamBox) -> ::Pair{Symbol, T}

Return a Pair of the name (with the index if available) and the value of the output variable of pb.

Quiqbox.outValCopy — Method

outValCopy(pb::ParamBox{T, V}) where {T} -> ParamBox{T, V, FLevel{0}}

Return a new ParamBox of which the input variable is a deep copy of the output variable of pb.

Quiqbox.outValOf — Method

outValOf(pb::ParamBox) -> Number

Return the value of the output variable of pb. Equivalent to pb().

Quiqbox.toggleDiff! — Method

toggleDiff!(pb::ParamBox) -> Bool

Toggle the differentiability of the input pb and then return it.

Quiqbox.add — Method

add(b1::CompositeGTBasisFuncs{T, D, <:Any, 1}, 
    b2::CompositeGTBasisFuncs{T, D, <:Any, 1}; roundDigits::Int=getAtolDigits(T)) where 
   {T, D} -> 
CompositeGTBasisFuncs{T, D, <:Any, 1}

Addition between two CompositeGTBasisFuncs{T, D, <:Any, 1}s such as BasisFunc and BasisFuncMix. roundDigits specifies the maximal number of digits after the radix point of the calculated values. The function can be called using + syntax with the keyword argument set to it default value.

≡≡≡ Example(s) ≡≡≡

julia> bf1 = genBasisFunc([1.,1.,1.], (2.,1.));

julia> bf2 = genBasisFunc([1.,1.,1.], (2.,2.));

julia> bf3 = bf1 + bf2;

julia> bf3.gauss[1].con() == bf1.gauss[1].con() + bf2.gauss[1].con()
true

Quiqbox.assignCenInVal! — Method

assignCenInVal!(center::AbstractVector{<:Real}, b::FloatingGTBasisFuncs{T, D}) -> 
SpatialPoint{T, D}

Change the input value of data stored in b.center (meaning the output value will also change according to the mapping function). Then, return the altered center.

Quiqbox.centerCoordOf — Method

centerCoordOf(bf::FloatingGTBasisFuncs{T}) where {T} -> Vector{T}

Return the center coordinate of the input FloatingGTBasisFuncs.

Quiqbox.centerOf — Method

centerOf(bf::FloatingGTBasisFuncs{T, D}) where {T, D} -> SpatialPoint{T, D}

Return the center of the input FloatingGTBasisFuncs.

Quiqbox.coordOf — Method

coordOf(sp::SpatialPoint{T}) where {T} -> Vector{T}

Get the coordinate represented by the input SpatialPoint.

Quiqbox.copyBasis — Function

copyBasis(b::GaussFunc, copyOutVal::Bool=true) -> GaussFunc

copyBasis(b::CompositeGTBasisFuncs, copyOutVal::Bool=true) -> CompositeGTBasisFuncs

Return a copy of the input basis. If copyOutVal is set to true, then only the output value(s) of the stored data will be copied, i.e., outValCopy is used to copy the ParamBoxs, otherwise inVarCopy is used.

≡≡≡ Example(s) ≡≡≡

julia> e = genExponent(3.0, x->x^2)
ParamBox{Float64, :α, FLevel{1}}(3.0)[∂][x_α]

julia> c = genContraction(2.0)
ParamBox{Float64, :d, FLevel{0}}(2.0)[∂][d]

julia> gf1 = GaussFunc(e, c);

julia> gf2 = copyBasis(gf1)
GaussFunc{Float64, FLevel{0}, FLevel{0}}(xpn()=9.0, con()=2.0, param)

julia> gf1.xpn() == gf2.xpn()
true

julia> (gf1.xpn[] |> gf1.xpn.map) == gf2.xpn[]
true

Quiqbox.decompose — Function

decompose(bf::CompositeGTBasisFuncs{T, D}, splitGaussFunc::Bool=false) -> 
Matrix{<:BasisFunc{T, D}}

Decompose a CompositeGTBasisFuncs into a Matrix of BasisFuncs. The sum of each column represents one orbital of the input basis function(s). If splitGaussFunc is true, then each column consists of the BasisFuncs with only 1 GaussFunc.

Quiqbox.dimOf — Method

dimOf(::DimensionalParamContainer) -> Int

Return the spatial dimension of the input parameterized container such as AbstractSpatialPoint and QuiqboxBasis.

Quiqbox.genBFuncsFromText — Method

genBFuncsFromText(content::String; adjustContent::Bool=false, 
                  adjustFunction::F=sciNotReplace, 
                  excludeFirstNlines::Int=0, excludeLastNlines::Int=0, 
                  center::Union{AbstractArray, 
                                NTuple{N, ParamBox}, 
                                Missing}=missing, 
                  unlinkCenter::Bool=false) where {N, F<:Function} -> 
Array{<:FloatingGTBasisFuncs, 1}

Generate a basis set from content which is either a basis set String in Gaussian format or the output from genBasisFuncText. For the former, adjustContent needs to be set to true. adjustFunction is only applied when adjustContent=true, which in default is a function used to detect and convert the format of the scientific notation in content.

excludeFirstNlines and excludeLastNlines are used to exclude first or last few lines of content if intended. center is used to assign a center coordinate for all the basis functions from content; when it's set to missing, it will try to read the center information in content, and leave the center as [NaN, NaN, Nan] if one cannot be found for each corresponding FloatingGTBasisFuncs. If unlinkCenter = true, the center of each FloatingGTBasisFuncs is a Base.deepcopy of the input center. Otherwise, they share the same underlying data so changing the value of one will affect others. If the center coordinate is included in content, it should be right above the subshell information for the FloatingGTBasisFuncs. E.g.:

"""
X        1.0                       0.0                       0.0                
S    1   1.0
         2.0                       1.0
"""

Quiqbox.genBasisFunc — Method

genBasisFunc(center::Union{AbstractVector{T}, Tuple{Vararg{T}}, SpatialPoint, Missing}, 
             args..., kws...) where {T<:Union{AbstractFloat, ParamBox}} -> 
Union{FloatingGTBasisFuncs{T}, Vector{<:FloatingGTBasisFuncs{T}}}

The constructor of FloatingGTBasisFuncs, but it also returns different kinds of collections (Vector) of them based on the input arguments. The first argument center specifies the center coordinate of the generated FloatingGTBasisFuncs, and can be left as missing for later assignment.

≡≡≡ Method 1 ≡≡≡

genBasisFunc(center, GsOrCoeffs, Ls; normalizeGTO=false) -> 
FloatingGTBasisFuncs

=== Positional argument(s) ===

GsOrCoeffs::Union{ AbstractGaussFunc{T1}, AbstractVector{<:AbstractGaussFunc{T1}}, Tuple{Vararg{AbstractGaussFunc{T1}}}, NTuple{2, T1}, NTuple{2, AbstractVector{T1}} } where {T1<:AbstractFloat}: A collection of concentric GaussFunc that will be used to construct the basis function. To simplify the procedure, it can also be in the form of a NTuple{2} of the exponent coefficient(s)::Union{AbstractFloat, AbstractVector{<:AbstractFloat}} and contraction coefficients::Union{AbstractFloat, AbstractVector{<:AbstractFloat}} of the GaussFunc(s) to be input.

Ls::Union{ T2, AbstractVector{T2}, NTuple{<:Any, T2} } where {T2<:Union{Tuple{Vararg{Int}}, LTuple}}: A collection of angular momentum(s) within the same subshell, in the Cartesian coordinate representation. E.g., for p shell it can be set to ((1,0,0), (0,1,0)). This will determine the number of spatial orbitals and their angular momentum respectively to be stored in the output FloatingGTBasisFuncs.

=== Keyword argument(s) ===

normalizeGTO::Bool: Determine whether the inside GaussFunc(s) will be normalized in the during the calculation.

=== Example(s) ===

julia> genBasisFunc([0.,0.,0.], GaussFunc(2.,1.), (0,1,0))
BasisFunc{Float64, 3, 1, 1, P3D{Float64, 0, 0, 0}}(center, gauss, l, normalizeGTO, param)[X⁰Y¹Z⁰][0.0, 0.0, 0.0]

≡≡≡ Method 2 ≡≡≡

genBasisFunc(center, GsOrCoeffs, subshell="S"; normalizeGTO=false) -> 
FloatingGTBasisFuncs

genBasisFunc(center, GsOrCoeffs, subshell, lFilter; normalizeGTO=false) -> 
FloatingGTBasisFuncs

=== Positional argument(s) ===

subshell::String: The third argument of the constructor can also be the name of a subshell, which will make sure the output is a BasisFuncs that contains the spatial orbitals that fully occupy the subshell.

lFilter::Tuple{Vararg{Bool}}: When this 4th argument is provided, it can determine the orbital(s) to be included based on the given subshell. The order of the corresponding orbital angular momentum(s) can be inspected using function orbitalLin.

=== Keyword argument(s) ===

normalizeGTO::Bool: Same as the one defined in method 1.

=== Example(s) ===

julia> genBasisFunc([0.,0.,0.], (2., 1.), "P")
BasisFuncs{Float64, 3, 1, 1, P3D{Float64, 0, 0, 0}, 3}(center, gauss, l, normalizeGTO, param)[3/3][0.0, 0.0, 0.0]

julia> genBasisFunc([0.,0.,0.], (2., 1.5), "P", (true, false, true))
BasisFuncs{Float64, 3, 1, 1, P3D{Float64, 0, 0, 0}, 2}(center, gauss, l, normalizeGTO, param)[2/3][0.0, 0.0, 0.0]

≡≡≡ Method 3 ≡≡≡

genBasisFunc(center, BSkey, atm="H"; unlinkCenter=false) -> 
Vector{<:FloatingGTBasisFuncs}

=== Positional argument(s) ===

BSkey::String: The name of an existed atomic basis set. The supported options are in ["STO-2G", "STO-3G", "STO-6G", "3-21G", "6-31G", "cc-pVDZ", "cc-pVTZ", "cc-pVQZ"].

atm::String: The name of the atom corresponding to the chosen basis set. The supported options are in ["H", "He", "Li", "Be", "B", "C", "N", "O", "F", "Ne", "Na", "Mg", "Al", "Si", "P", "S", "Cl", "Ar", "K", "Ca"].

=== Keyword argument(s) ===

unlinkCenter::Bool: Determine whether the centers of constructed FloatingGTBasisFuncs are linked to each other. If set to true, the center of each FloatingGTBasisFuncs is a Base.deepcopy of each other. Otherwise, they share the same underlying data so changing the value of one will affect others.

=== Example(s) ===

julia> genBasisFunc([0.,0.,0.], "6-31G");

julia> genBasisFunc([0.,0.,0.], "STO-3G", "Li");

≡≡≡ Method 4 ≡≡≡

genBasisFunc(b::FloatingGTBasisFuncs{T, D}, newFieldVal) where {T, D} -> 
FloatingGTBasisFuncs{T, D}

=== Positional argument(s) ===

field::Union{ SpatialPoint{T, D}, Tuple{Vararg{AbstractGaussFunc{T}}}, Tuple{Vararg{LTuple{D, 𝑙}}} where 𝑙, Bool } where {T<:AbstractFloat, D}: Any one of the fields inside a FloatingGTBasisFuncs except param.

This method outputs a FloatingGTBasisFuncs that has identical fields as the input one except the field that can be replaced by newFieldVal (and param if the replaced field contains ParamBox).

Quiqbox.genBasisFuncText — Method

genBasisFuncText(bf::FloatingGTBasisFuncs; 
                 norm::Real=1.0, printCenter::Bool=true, roundDigits::Int=-1) -> String

Generate the text of input FloatingGTBasisFuncs. norm is the additional normalization factor. If printCenter is true, the center coordinate will be added to the first line of the output String. roundDigits specifies the rounding digits for the parameters inside bf. When set to negative, no rounding will be performed.

Quiqbox.genBasisFuncText — Method

genBasisFuncText(bs::Union{AbstractVector{<:FloatingGTBasisFuncs{T, D}}, 
                           Tuple{Vararg{FloatingGTBasisFuncs{T, D}}}}; 
                 norm::Real=1.0, printCenter::Bool=true, 
                 groupCenters::Bool=true, roundDigits::Int=-1) where {T, D} -> 
String

Generate the text of input basis set (consisting of FloatingGTBasisFuncs). norm is the additional normalization factor. groupCenters determines whether the function will group the basis functions with same center together.

Quiqbox.genContraction — Method

genContraction(c::T, mapFunction::Function; canDiff::Bool=true, 
               dataName::Symbol=:undef) where {T<:AbstractFloat} -> 
ParamBox{T, :d}

genContraction(c::Array{T, 0}, mapFunction::Function; canDiff::Bool=true, 
               dataName::Symbol=:undef) where {T<:AbstractFloat} -> 
ParamBox{T, :d}

Construct a ParamBox for an contraction coefficient given a value. Keywords mapFunction and canDiff work the same way as in a general constructor of a ParamBox.

Quiqbox.genContraction — Method

genContraction(c::T) where {T<:AbstractFloat} -> ParamBox{T, :d}

genContraction(c::Array{T, 0}) where {T<:AbstractFloat} -> ParamBox{T, :d}

Quiqbox.genContraction — Method

genContraction(pb::ParamBox{T}) where {T<:AbstractFloat} -> ParamBox{T, :d}

Convert a ParamBox to an exponent coefficient parameter.

Quiqbox.genExponent — Method

genExponent(e::T, mapFunction::Function; canDiff::Bool=true, 
            dataName::Symbol=:undef) where {T<:AbstractFloat} -> 
ParamBox{T, :α}

genExponent(e::Array{T, 0}, mapFunction::Function; canDiff::Bool=true, 
            dataName::Symbol=:undef) where {T<:AbstractFloat} -> 
ParamBox{T, :α}

Construct a ParamBox for an exponent coefficient given a value. Keywords mapFunction and canDiff work the same way as in a general constructor of a ParamBox.

Quiqbox.genExponent — Method

genExponent(e::T) where {T<:AbstractFloat} -> ParamBox{T, :α}

genExponent(e::Array{T, 0}) where {T<:AbstractFloat} -> ParamBox{T, :α}

Quiqbox.genExponent — Method

genExponent(pb::ParamBox{T}) where {T<:AbstractFloat} -> ParamBox{T, :α}

Convert a ParamBox to the container of an exponent coefficient.

Quiqbox.genSpatialPoint — Method

genSpatialPoint(comp::T, index::Int) where {T<:AbstractFloat} -> ParamBox{T}

genSpatialPoint(comp::Array{T, 0}, index::Int) where {T<:AbstractFloat} -> ParamBox{T}

genSpatialPoint(comp::T, index::Int, mapFunction::Function; canDiff::Bool=true, 
                dataName::Symbol=:undef) where {T<:AbstractFloat} -> 
ParamBox{T}

genSpatialPoint(comp::Array{T, 0}, index::Int, mapFunction::Function; 
                canDiff::Bool=true, dataName::Symbol=:undef) where 
               {T<:AbstractFloat} -> 
ParamBox{T}

genSpatialPoint(comp::ParamBox{T}, index::Int) where {T<:AbstractFloat} -> ParamBox{T}

Return the component of a SpatialPoint given its value (or 0-D container) and index.

≡≡≡ Example(s) ≡≡≡

julia> genSpatialPoint(1.2, 1)
ParamBox{Float64, :X, FLevel{0}}(1.2)[∂][X]

julia> pointY1 = fill(2.0);

julia> Y1 = genSpatialPoint(pointY1, 2)
ParamBox{Float64, :Y, FLevel{0}}(2.0)[∂][Y]

julia> pointY1[] = 1.5;

julia> Y1
ParamBox{Float64, :Y, FLevel{0}}(1.5)[∂][Y]

Quiqbox.genSpatialPoint — Method

genSpatialPoint(point::Union{Tuple{Vararg{AbstractFloat}}, 
                                   AbstractVector{<:AbstractFloat}}, 
                mapFunction::F=itself; canDiff::Bool=true, dataName::Symbol=:undef) -> 
SpatialPoint

genSpatialPoint(point::Union{Tuple{Vararg{Array{<:AbstractFloat, 0}}}, 
                             AbstractVector{<:Array{<:AbstractFloat, 0}}}, 
                mapFunction::F=itself; canDiff::Bool=true, dataName::Symbol=:undef) -> 
SpatialPoint

The constructor of SpatialPoint. Keywords mapFunction and canDiff work the same way as in a general constructor of a ParamBox. If roundDigits < 0 or point is a 0-dimensional Array, there won't be rounding for input data.

≡≡≡ Example(s) ≡≡≡

julia> v1 = [1.0, 2.0, 3.0];

julia> genSpatialPoint(v1)
SpatialPoint{Float64, 3, P3D{Float64, 0, 0, 0}}(param)[1.0, 2.0, 3.0][∂][∂][∂]

julia> v2 = [fill(1.0), 2.0, 3.0];

julia> p2 = genSpatialPoint(v2); p2[1]
ParamBox{Float64, :X, FLevel{0}}(1.0)[∂][X]

julia> v2[1][] = 1.2
1.2

julia> p2[1]
ParamBox{Float64, :X, FLevel{0}}(1.2)[∂][X]

Quiqbox.genSpatialPoint — Method

genSpatialPoint(point::Union{Tuple{Vararg{ParamBox}}, AbstractVector{<:ParamBox}}) -> 
SpatialPoint

Convert a collection of ParamBoxs to a SpatialPoint.

Quiqbox.getParams — Function

getParams(pbc::ParamBox, symbol::Union{Symbol, Missing}=missing; 
          forDifferentiation::Bool=false) -> 
Union{ParamBox, Nothing}

getParams(pbc::ParameterizedContainer, symbol::Union{Symbol, Missing}=missing; 
          forDifferentiation::Bool=false) -> 
AbstractVector{<:ParamBox}

getParams(pbc::Union{AbstractArray, Tuple}, symbol::Union{Symbol, Missing}=missing; 
          forDifferentiation::Bool=false) -> 
AbstractVector{<:ParamBox}

Return the parameter(s) stored in the input container. If symbol is set to missing, then return all parameter(s); if it's set to the Symbol of a parameter (e.g., :α₁ will match any pb::ParamBox such that getVar(pb) == :α₁; :α will match all the pbs that are ParamBox{<:Any, α}. forDifferentiation determines whether searching through the Symbol(s) of the independent variable(s) represented by pbc during the differentiation process. If the first argument is a collection, its entries must be ParamBox containers.

Quiqbox.lOf — Method

lOf(::FloatingGTBasisFuncs) -> Int

Return the total orbital angular momentum quantum number (in Cartesian coordinate representation).

Quiqbox.markParams! — Method

markParams!(b::Union{AbstractVector{T}, T, Tuple{Vararg{T}}}, 
            filterMapping::Bool=false) where {T<:ParameterizedContainer} -> 
Vector{<:ParamBox}

Mark the parameters (ParamBox) in b. The parameters that will be considered identical in the differentiation procedure will be marked with same index. filterMapping determines whether filtering out (i.e. not return) the extra ParamBoxs that have the same indices despite may having different mapping functions.

Quiqbox.mul — Method

mul(a1::Real, a2::CompositeGTBasisFuncs{T, D, <:Any, 1}; 
    normalizeGTO::Union{Bool, Missing}=missing, 
    roundDigits::Int=getAtolDigits(T)) where {T, D} -> 
CompositeGTBasisFuncs{T, D, <:Any, 1}

mul(a1::CompositeGTBasisFuncs{T, D, <:Any, 1}, a2::Real; 
    normalizeGTO::Union{Bool, Missing}=missing, 
    roundDigits::Int=getAtolDigits(T)) where {T, D} -> 
CompositeGTBasisFuncs{T, D, <:Any, 1}

mul(a1::CompositeGTBasisFuncs{T, D, <:Any, 1}, 
    a2::CompositeGTBasisFuncs{T, D, <:Any, 1}; 
    normalizeGTO::Union{Bool, Missing}=missing, 
    roundDigits::Int=getAtolDigits(T)) where {T, D} -> 
CompositeGTBasisFuncs{T, D, <:Any, 1}

Multiplication between two CompositeGTBasisFuncs{T, D, <:Any, 1}s (e.g., BasisFunc and BasisFuncMix), or a Real number and a CompositeGTBasisFuncs{T, D, <:Any, 1}. If normalizeGTO is set to missing (in default), The GaussFunc inside the output containers will be normalized only if all the input bases hold .normalizeGTO == true. roundDigits specifies the maximal number of digits after the radix point of the calculated values. When set to negative, no rounding will be performed. The function can be called using * syntax with the keyword arguments set to their default values.

≡≡≡ Example(s) ≡≡≡

julia> bf1 = genBasisFunc([1.0, 1.0, 1.0], ([2.0, 1.0], [0.1, 0.2]))
BasisFunc{Float64, 3, 0, 2, P3D{Float64, 0, 0, 0}}(center, gauss, l, normalizeGTO, param)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

julia> bf2 = bf1 * 2
BasisFunc{Float64, 3, 0, 2, P3D{Float64, 0, 0, 0}}(center, gauss, l, normalizeGTO, param)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

julia> getindex.(getproperty.(bf2.gauss, :con))
(0.2, 0.4)

julia> bf3 = bf1 * bf2
BasisFunc{Float64, 3, 0, 3, P3D{Float64, 0, 0, 0}}(center, gauss, l, normalizeGTO, param)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

Quiqbox.mul — Method

mul(gf::GaussFunc{T}, coeff::Real; roundDigits::Int=getAtolDigits(T)) where {T} -> 
GaussFunc

mul(coeff::Real, gf::GaussFunc{T}; roundDigits::Int=getAtolDigits(T)) where {T} -> 
GaussFunc

mul(gf1::GaussFunc{T}, gf2::GaussFunc{T}; 
    roundDigits::Int=getAtolDigits(T)) where {T} -> 
GaussFunc

Multiplication between a Real number and a GaussFunc or two GaussFuncs. roundDigits specifies the maximal number of digits after the radix point of the calculated values. When set to negative, no rounding will be performed. The function can be called using * syntax with the keyword argument set to it default value.

≡≡≡ Example(s) ≡≡≡

julia> gf1 = GaussFunc(3.0, 1.0)
GaussFunc{Float64, FLevel{0}, FLevel{0}}(xpn()=3.0, con()=1.0, param)

julia> gf1 * 2
GaussFunc{Float64, FLevel{0}, FLevel{0}}(xpn()=3.0, con()=2.0, param)

julia> gf1 * gf1
GaussFunc{Float64, FLevel{0}, FLevel{0}}(xpn()=6.0, con()=1.0, param)

julia> gf1 * 2 * gf1
GaussFunc{Float64, FLevel{0}, FLevel{0}}(xpn()=6.0, con()=2.0, param)

Quiqbox.orbitalNumOf — Function

orbitalNumOf(subshell::String, D::Integer=3) -> Int

Return the size (number of orbitals) of each subshell in D dimensional real space.

Quiqbox.orbitalNumOf — Method

orbitalNumOf(b::QuiqboxBasis) -> Int

Return the numbers of orbitals of the input basis.

Quiqbox.shift — Method

shift(bf::FloatingGTBasisFuncs{T, D, 𝑙, GN, PT, 1}, 
      dl::Union{Vector{Int}, NTuple{D, Int}, LTuple{D}}, op::Function=+) where 
     {T, D, 𝑙, GN, PT} -> 
BasisFunc{T, D, <:Any, GN, PT}

Shift (+ as the default binary operator op) the angular momentum (in Cartesian representation) of the input FloatingGTBasisFuncs given dl that specifies the change of each component.

Quiqbox.sortBasis — Method

sortBasis(b::GTBasis{T, D}; roundDigits::Int=getAtolDigits(T)) where {T, D} -> 
GTBasis{T, D}

Reconstruct a GTBasis by sorting the GTBasisFuncs stored in the input one.

Quiqbox.sortBasis — Method

sortBasis(bs::Union{AbstractArray{<:CompositeGTBasisFuncs{T, D}}, 
                    Tuple{Vararg{CompositeGTBasisFuncs{T, D}}}}; 
          roundDigits::Int=getAtolDigits(T)) where {T, D} -> 
Vector{<:CompositeGTBasisFuncs{T, D}}

Sort basis functions. roundDigits specifies the rounding digits for the parameters stored in each CompositeGTBasisFuncs when comparing them. When set to negative, no rounding will be performed.

Quiqbox.sortBasis — Method

sortBasis(bs::Tuple{Vararg{CompositeGTBasisFuncs{T, D}}}; 
          roundDigits::Int=getAtolDigits(T)) where {T, D} -> 
Tuple{Vararg{CompositeGTBasisFuncs{T, D}}}

Quiqbox.sortBasisFuncs — Method

sortBasisFuncs(bs::AbstractArray{<:FloatingGTBasisFuncs{T, D}}, 
               groupCenters::Bool=false; roundDigits::Int=getAtolDigits(T)) where 
              {T, D} -> 
Vector

Sort FloatingGTBasisFuncs. If groupCenters = true, Then the function will return an Vector{<:Vector{<:FloatingGTBasisFuncs}} in which the elements are grouped basis functions with same center coordinates. roundDigits specifies the rounding digits for center coordinates when comparing them. When set to negative, no rounding will be performed.

Quiqbox.sortBasisFuncs — Method

sortBasisFuncs(bs::Tuple{Vararg{FloatingGTBasisFuncs{T, D}}}, groupCenters::Bool=false; 
               roundDigits::Int=getAtolDigits(T)) where {T, D} -> 
Tuple

Quiqbox.sortPermBasis — Method

sortPermBasis(bs::AbstractArray{<:CompositeGTBasisFuncs{T, D}}; 
              roundDigits::Int=getAtolDigits(T)) where {T, D} -> 
Vector{Int}

Return a Vector of indices I such that bs[I] ==sortBasis (bs; roundDigits)[I].

Quiqbox.sortPermBasisFuncs — Method

sortPermBasisFuncs(bs::Union{AbstractArray{<:FloatingGTBasisFuncs{T, D}}, 
                             Tuple{Vararg{FloatingGTBasisFuncs{T, D}}}}) where 
                  {T, D} -> 
Vector{Int}

Return a Vector of indices I such that bs[I] ==sortBasisFuncs (bs; roundDigits)[I].

Quiqbox.subshellOf — Method

subshellOf(::FloatingGTBasisFuncs) -> String

Return the corresponding subshell of the input FloatingGTBasisFuncs.

Quiqbox.runHF — Method

runHF(bs, nuc, nucCoords, config=HFconfig(), N=getCharge(nuc); printInfo=true) -> 
HFfinalVars

runHF(bs, nuc, nucCoords, N=getCharge(nuc), config=HFconfig(); printInfo=true) -> 
HFfinalVars

Main function to run Hartree-Fock in Quiqbox.

≡≡≡ Positional argument(s) ≡≡≡

bs::Union{ BasisSetData{T, D}, AbstractVector{<:AbstractGTBasisFuncs{T, D}}, Tuple{Vararg{AbstractGTBasisFuncs{T, D}}} } where {T, D}: The basis set used for the Hartree-Fock approximation.

nuc::Union{ NTuple{NN, String} where NN, AbstractVector{String} }: The nuclei in the studied system.

nucCoords::Union{AbstractArray{Tuple{Vararg{T, D}}, 1}, Tuple{Vararg{AbstractVector{T}}}, Tuple{Vararg{Tuple{Vararg{T, D}}, NN}}, AbstractVector{VT} where VT<:AbstractVector{T}} where {T, D, NN}: The coordinates of corresponding nuclei.

config::HFconfig: The Configuration of selected Hartree-Fock method. For more information please refer to HFconfig.

N::Union{Int, Tuple{Int}, NTuple{2, Int}}: Total number of electrons, or the number(s) of electrons with same spin configurations(s).

≡≡≡ Keyword argument(s) ≡≡≡

printInfo::Bool: Whether print out the information of iteration steps and result.

Quiqbox.runHFcore — Method

runHFcore(HTtype, scfConfig, N, Hcore, HeeI, S, X, C0, 
          printInfo=false, maxStep=1000, earlyStop=true) -> 
Tuple{Tuple{Vararg{HFtempVars}}, Bool}

The core function of runHF which returns the data collected during the iteration and the result of whether the SCF procedure is converged.

=== Positional argument(s) ===

HTtype::Val{HFT} where HFT: Hartree-Fock method type. Available values of HFT are :RHF, :UHF.

scfConfig::SCFconfig: The SCF iteration configuration.

N::NTuple{HFTS, Int} where HFTS: The numbers of electrons with same spin configurations.

Hcore::AbstractMatrix{T} where T: The core Hamiltonian of the electronic Hamiltonian.

HeeI::AbstractArray{T, 4} where T: The electron-electron interaction tensor (in the chemists' notation) which includes both the Coulomb interactions and the Exchange Correlations.

S::AbstractMatrix{T} where T: The overlap matrix of the used basis set.

X::AbstractMatrix{T} where T: The transformation matrix of S.

C0::NTuple{HFTS, AbstractMatrix{T}} where {HFTS, T}: Initial guess of the coefficient matrix(s) of the canonical spin-orbitals.

printInfo::Bool: Whether print out the information of iteration steps and result.

maxStep::Int: Maximum iteration steps allowed regardless if the iteration converges.

earlyStop::Bool: Whether automatically terminate (or skip) a convergence method early when its performance becomes unstable or poor.

Quiqbox.gradOfHFenergy — Method

gradOfHFenergy(par, basis, C, nuc, nucCoords, N=getCharge(nuc)) ->
AbstractVector

gradOfHFenergy(par, bs, S, C, nuc, nucCoords, N=getCharge(nuc)) ->
AbstractVector

Two methods of gradOfHFenergy.

≡≡≡ Positional argument(s) ≡≡≡

par::AbstractVector{<:ParamBox}: The parameters for differentiation.

basis::GTBasis{T, D} where {T, D}: Basis set information.

C::NTuple{<:Any, AbstractMatrix{T}} where T: The coefficient matrix(s) of the canonical orbitals with respect to the selected basis set.

nuc::Union{ NTuple{NN, String} where NN, AbstractVector{String} }: The nuclei in the studied system.

nucCoords::Union{AbstractArray{Tuple{Vararg{T, D}}, 1}, Tuple{Vararg{AbstractVector{T}}}, Tuple{Vararg{Tuple{Vararg{T, D}}, NN}}, AbstractVector{VT} where VT<:AbstractVector{T}} where {T, D, NN}: The coordinates of corresponding nuclei.

N::Union{Int, Tuple{Int}, NTuple{2, Int}}: Total number of electrons, or the number(s) of electrons with same spin configurations(s).

bs::Union{ NTuple{BN, GTBasisFuncs{T, D, 1}}, AbstractVector{<:GTBasisFuncs{T, D, 1}} } where {T, D}: A collection of basis functions.

S::AbstractMatrix{T} where T: The overlap lap of the basis set when bs is provided as the second argument.

NOTE 1: If any of these two methods is applied, the user needs to make sure the row orders as well as the colum orders of C and (or) S are consistent with the element order of bs (basis.basis). ``

Quiqbox.gridCoordOf — Method

gridCoordOf(gb::GridBox{T}) where {T} -> Tuple{Vararg{Vector{T}}}

Return the coordinates of the grid points stored in gb.

Quiqbox.gradDescent! — Method

gradDescent!(vars::AbstractVector{T}, grad::AbstractVector{T}, η::T=T(0.001), 
             threshold::T=2sqrt(length(grad))/(25norm(η))) where {T} -> 
vars::AbstractVector{T}

Default gradient descent (GD) method used in optimizeParams!. vars are the input variables of a function with corresponding gradient grad. η is the learning rate (step size) of the gradient descent. threshold is the clipping threshold of grad which will be renormalized if it's larger then threshold to prevent gradient exploding. gradDescent! modifies the vars and returns the updated value. It can be replaced by a more advanced GD function through customizing POconfig.

Quiqbox.optimizeParams! — Method

optimizeParams!(pbs, bs, nuc, nucCoords, 
                config=POconfig(), N=getCharge(nuc); printInfo=true) -> 
Tuple{Vector{T}, Matrix{T}, Matrix{T}} where {T}

optimizeParams!(pbs, bs, nuc, nucCoords, 
                N=getCharge(nuc), config=POconfig(); printInfo=true) -> 
Tuple{Vector{T}, Matrix{T}, Matrix{T}} where {T}

The main function to optimize the parameters of a given basis set. It returns a Tuple of the energies, the parameter values and the gradients of all the steps. For latter two, each column is the result of each step.

=== Positional argument(s) ===

pbs::AbstractVector{<:ParamBox{T}}: The parameters to be optimized that are extracted from the basis set. If the parameter is marked as "differentiable", the value of its input variable will be optimized.

bs::Union{ Tuple{Vararg{AbstractGTBasisFuncs{T, D}}}, AbstractVector{<:AbstractGTBasisFuncs{T, D}} }: The basis set to be optimized.

nuc::Union{ NTuple{NN, String} where NN, AbstractVector{String} }: The nuclei in the studied system.

nucCoords::Union{AbstractArray{Tuple{Vararg{T, D}}, 1}, Tuple{Vararg{AbstractVector{T}}}, Tuple{Vararg{Tuple{Vararg{T, D}}, NN}}, AbstractVector{VT} where VT<:AbstractVector{T}} where {T, D, NN}: The coordinates of corresponding nuclei.

config::POconfig: The Configuration of selected parameter optimization method. For more information please refer to POconfig.

N::Union{Int, Tuple{Int}, NTuple{2, Int}}: Total number of electrons, or the number(s) of electrons with same spin configurations(s).

=== Keyword argument(s) ===

printInfo::Bool: Whether print out the information of iteration steps.

Quiqbox.changeHbasis — Method

changeHbasis(HFres::HFfinalVars) -> NTuple{2, Any}

Return the one-body and two-body integrals on the basis of the canonical orbitals using the result of a Hartree-Fock method HFres.

Quiqbox.changeHbasis — Method

changeHbasis(b::GTBasis{T, D}, 
             nuc::NTuple{NN, String}, nucCoords::NTuple{NN, NTuple{D, T}}, 
             C::Union{AbstractMatrix{T}, NTuple{2, AbstractMatrix{T}}}) where 
            {T, D, NN} -> 
NTuple{2, Any}

Return the one-body and two-body integrals after a change of basis based on the input C, given the basis set information b. The type of each element in the returned Tuple is consistent with the cases where the first argument of changeHbasis is an AbstractArray.

Quiqbox.changeHbasis — Method

changeHbasis(twoBodyInt::AbstractArray{T, 4}, 
             C1::AbstractMatrix{T}, C2::AbstractMatrix{T}) where {T} -> 
AbstractArray{T, 4}

Change the basis of the input two-body integrals twoBodyInt based on two orbital coefficient matrices C1 and C2 for different spin configurations (e.g., the unrestricted case). The output is a 3-element Tuple of which the first 2 elements are the spatial integrals of each spin configurations respectively, while the last element is the Coulomb interactions between orbitals with different spins.

Quiqbox.changeHbasis — Method

changeHbasis(DbodyInt::AbstractArray{T, D}, C::AbstractMatrix{T}) where {T} -> 
AbstractArray{T, D}

Change the basis of the input one-body / two-body integrals DbodyInt based on the orbital coefficient matrix C.

Quiqbox.getCanOrbitals — Method

getCanOrbitals(fVars::HFfinalVars{T, D, <:Any, NN}) where {T, D, NN} -> 
CanOrbital{T, D, NN}

Generate a set of canonical orbitals from the result of a Hartree-Fock approximation.

Quiqbox.nnRepulsions — Method

nnRepulsions(nuc::Union{NTuple{NN, String}, AbstractVector{String}}, 
             nucCoords::Union{AbstractArray{Tuple{Vararg{T, D}}, 1}, Tuple{Vararg{AbstractVector{T}}}, Tuple{Vararg{Tuple{Vararg{T, D}}, NN}}, AbstractVector{VT} where VT<:AbstractVector{T}}) where {NN, D, T} -> 
T

Return the nuclear repulsion energy given nuclei nuc and their coordinates nucCoords.

Quiqbox.coreH — Method

coreH(bs::Union{GTBasis, Tuple{Vararg{AbstractGTBasisFuncs{T, D}}}, 
                AbstractVector{<:AbstractGTBasisFuncs{T, D}}}, 
      nuc::Union{NTuple{NN, String}, AbstractVector{String}}, 
      nucCoords::Union{AbstractArray{Tuple{Vararg{T, D}}, 1}, Tuple{Vararg{AbstractVector{T}}}, Tuple{Vararg{Tuple{Vararg{T, D}}, NN}}, AbstractVector{VT} where VT<:AbstractVector{T}}) where {T, D, NN} -> 
Matrix{T}

Return the core Hamiltonian given a basis set and the corresponding nuclei with their coordinates (in atomic units).

Quiqbox.coreHij — Method

coreHij(bf1::AbstractGTBasisFuncs{T, D, 1}, bf2::AbstractGTBasisFuncs{T, D, 1}, 
        nuc::Union{NTuple{NN, String}, AbstractVector{String}}, 
        nucCoords::Union{AbstractArray{Tuple{Vararg{T, D}}, 1}, Tuple{Vararg{AbstractVector{T}}}, Tuple{Vararg{Tuple{Vararg{T, D}}, NN}}, AbstractVector{VT} where VT<:AbstractVector{T}}) where {T, D, NN} -> 
T

Return a matrix element of the core Hamiltonian given two basis functions.

Quiqbox.eKinetic — Method

eKinetic(bf1::AbstractGTBasisFuncs{T, D, 1}, bf2::AbstractGTBasisFuncs{T, D, 1}) where 
        {T, D} -> 
T

Return the electron kinetic energy between two basis functions.

Quiqbox.eKinetics — Method

eKinetics(bs::Union{Tuple{Vararg{AbstractGTBasisFuncs{T, D}}}, 
                    AbstractVector{<:AbstractGTBasisFuncs{T, D}}}) where {T, D} -> 
Matrix{T}

Return the electron kinetic energy matrix given a basis set.

Quiqbox.neAttraction — Method

neAttraction(bf1::AbstractGTBasisFuncs{T, D, 1}, bf2::AbstractGTBasisFuncs{T, D, 1}, 
             nuc::Union{NTuple{NN, String}, AbstractVector{String}}, 
             nucCoords::Union{AbstractArray{Tuple{Vararg{T, D}}, 1}, Tuple{Vararg{AbstractVector{T}}}, Tuple{Vararg{Tuple{Vararg{T, D}}, NN}}, AbstractVector{VT} where VT<:AbstractVector{T}}) where {T, D, NN} -> 
T

Return the nuclear attraction between two basis functions, provided with the nuclei and their coordinates (in the atomic units).

Quiqbox.neAttractions — Method

neAttractions(bs::Union{Tuple{Vararg{AbstractGTBasisFuncs{T, D}}}, 
                        AbstractVector{<:AbstractGTBasisFuncs{T, D}}}, 
              nuc::Union{NTuple{NN, String}, AbstractVector{String}}, 
              nucCoords::Union{AbstractArray{Tuple{Vararg{T, D}}, 1}, Tuple{Vararg{AbstractVector{T}}}, Tuple{Vararg{Tuple{Vararg{T, D}}, NN}}, AbstractVector{VT} where VT<:AbstractVector{T}}) where {T, D, NN} -> 
Matrix{T}

Return the nuclear attraction matrix given a basis set and the corresponding nuclei with their coordinates (in atomic units).

Quiqbox.overlap — Method

overlap(bf1::AbstractGTBasisFuncs{T, D, 1}, bf2::AbstractGTBasisFuncs{T, D, 1}) where 
       {T, D, 1} -> 
T

Return the orbital overlap between two basis functions.

Quiqbox.overlaps — Method

overlaps(bs::Union{Tuple{Vararg{AbstractGTBasisFuncs{T, D}}}, 
                   AbstractVector{<:AbstractGTBasisFuncs{T, D}}}) where {T, D} -> 
Matrix{T}

Return the orbital overlap matrix given a basis set.

Quiqbox.eeInteraction — Method

eeInteraction(bf1::AbstractGTBasisFuncs{T, D, 1}, 
              bf2::AbstractGTBasisFuncs{T, D, 1}, 
              bf3::AbstractGTBasisFuncs{T, D, 1}, 
              bf4::AbstractGTBasisFuncs{T, D, 1}) where {T, D} -> 
T

Return an electron-electron interaction tensor element given four basis functions (ordered in the chemists' notation).

Quiqbox.eeInteractions — Method

eeInteractions(bs::Union{Tuple{Vararg{AbstractGTBasisFuncs{T, D}}}, 
                         AbstractVector{<:AbstractGTBasisFuncs{T, D}}}) -> 
Array{T, 4}

Return the tensor of electron-electron interactions (in the chemists' notation) given a basis set.

Quiqbox.eeIuniqueIndicesOf — Method

eeIuniqueIndicesOf(basisSetSize::Int) -> Vector{Vector{Int}}

Return the unique matrix element indices (in the chemists' notation) of electron-electron interactions given the size of a basis set.