Core Types

Quiqbox.ParamBox — Type

ParamBox{V, T} <: DifferentiableParameter{ParamBox, T}

Parameter container that enables parameter differentiations.

≡≡≡ Field(s) ≡≡≡

data::T: Stored parameter. It can be accessed through syntax [].

canDiff::Bool: Indicator that whether this container should be marked as differentiable.

≡≡≡ Initialization Method(s) ≡≡≡

ParamBox(data::Number, name::Symbol=:undef; mapFunction::Function=itself, 
         canDiff::Bool=true, paramType::Type{T}=Float64) -> 
ParamBox{T}

name specifies the name of the variable to be stored, which helps with symbolic representation and automatic differentiation.

mapFunction is for the case to the store the variable that is a dependent variable (math function) f(x) of another variable x which is the actually stored in the struct, and linked to the f(x) via the mapFunction. After initializing the ParamBox, e.g pb1 = ParamBox(x, mapFunction=f), pb.data[] returns x, and pb.data() returns f(x).

canDiff is used to mark the (independent) variable as differentiable when set to true, otherwise the ParamBox will be ignored in any differentiation process like a constant.

paramType specifies the type of the stored variable to avoid data type mutation.

≡≡≡ Example(s) ≡≡≡

julia> Quiqbox.ParamBox(1.0)
ParamBox{Float64}(1.0)[∂]

NOTE: When the parameter inside x::ParamBox is marked as "differentiable" (a.k.a. x.canDiff=true), "[∂]" in the printing info is in color green, otherwise it's in grey.

Quiqbox.GaussFunc — Type

GaussFunc <: AbstractGaussFunc

A single contracted gaussian function struct from package Quiqbox.

≡≡≡ Field(s) ≡≡≡

xpn::ParamBox{:𝛼, Float64}：Exponent of the gaussian function.

con::ParamBox{:𝑑, Float64}: Contraction coefficient of the gaussian function.

param::NTuple{2, ParamBox}: A Tuple that stores the ParamBoxs of xpn and con.

≡≡≡ Initialization Method(s) ≡≡≡

GaussFunc(xpn::ParamBox{:α, Float64}, 
          con::ParamBox{:d, Float64}) -> 
GaussFunc

GaussFunc(xpn::Real, con::Real) -> GaussFunc

Quiqbox.BasisFunc — Type

BasisFunc{S, GN} <: FloatingGTBasisFuncs{S, GN, 1}

A (floating) basis function with the center attached to it instead of any nucleus.

≡≡≡ Field(s) ≡≡≡

center::NTuple{3, ParamBox}: The center coordinate in form of a 3-element ParamBox-type Tuple.

gauss::NTuple{N, GaussFunc}: Gaussian functions within the basis function.

subshell::String: The subshell (angular momentum symbol).

ijk::Tuple{String}: Cartesian representation (pseudo-quantum number) of the angular momentum orientation. E.g., s would be ("X⁰Y⁰Z⁰")

normalizeGTO::Bool: Whether the GTO::GaussFunc will be normalized in calculations.

param::Tuple{Vararg{<:ParamBox}}： All the tunable parameters::ParamBox stored in the BasisFunc.

≡≡≡ Initialization Method(s) ≡≡≡

BasisFunc(center::Tuple{Vararg{<:ParamBox}}, gauss::NTuple{GN, GaussFunc}, 
          ijk::Array{Int, 1}, normalizeGTO::Bool) where {GN} -> 
BasisFunc{S, GN}

BasisFunc(cen::Tuple{Vararg{<:ParamBox}}, gauss::GaussFunc, ijk::Vector{Int}, 
          normalizeGTO::Bool) ->
BasisFunc{S, 1}

Quiqbox.BasisFuncs — Type

BasisFuncs{S, GN, ON} <: FloatingGTBasisFuncs{S, GN, ON}

A group of basis functions with identical parameters except they have different subshell under the specified angular momentum. It has the same fields as BasisFunc and specifically, for ijk, instead of being a 1-element Tuple, the size of the Tuple is the size of the corresponding subshell.

Quiqbox.GTBasis — Type

GTBasis{N, BT} <: BasisSetData{N}

The container to store basis set information.

≡≡≡ Field(s) ≡≡≡

basis::Array{<:AbstractGTBasisFuncs, 1}: Basis set.

S::Array{<:Number, 2}: Overlap matrix.

Te::Array{<:Number, 2}: Kinetic energy part of the electronic core Hamiltonian.

eeI::Array{<:Number, 4}: Electron-electron interaction.

getVne::Function: A Function that returns the nuclear attraction Hamiltonian when nuclei::Array{String, 1} and their coordinates::Array{<:AbstractArray, 1} are input.

getHcore::Function: Similar as getVne, a Function that returns the core Hamiltonian when nuclei and their coordinates of same DataType are input.

≡≡≡ Initialization Method(s) ≡≡≡

GTBasis(basis::Vector{<:AbstractGTBasisFuncs}, S::Matrix{<:Number}, 
        Te::Matrix{<:Number}, eeI::Array{<:Number, 4}) -> 
GTBasis

Quiqbox.GridBox — Type

GridBox{NX, NY, NZ} <: SemiMutableParameter{GridBox, Float64}

A struct that stores coordinates of grid points in terms of both Vectors and ParamBoxs.

≡≡≡ Field(s) ≡≡≡

num::Int: Total number of the grid points.

spacing::Float64: The length between adjacent grid points.

box::Vector{NTuple{3, ParamBox}}: The coordinates of grid points in terms of ParamBoxs.

coord::Array{Array{Float64, 1}, 1}: The coordinates of grid points in terms of Vectors.

≡≡≡ Initialization Method(s) ≡≡≡

GridBox(nGrids::NTuple{3, Int}, spacing::Real=10, 
        centerCoord::Array{<:Real, 1}=[0.0,0.0,0.0];
        canDiff::Bool=true, index::Int=0) -> GridBox

Constructor of a general GridBox that doesn't have to shape as a cube. nGrid is a 3-element Tuple that specifies the number of grids (number of grid points - 1) along 3 dimensions. spacing specifies the length between adjacent grid points. centerCoord specifies the geometry center coordinate of the box. canDiff determines whether the ParamBox should be marked as differentiable. index defines the index number for the actual parameter: spacing L, with the default value 0 it would be L₀.

Quiqbox.SCFconfig — Type

SCFconfig{N} <: ImmutableParameter{SCFconfig, Any}

The struct for SCF iteration configurations.

≡≡≡ Field(s) ≡≡≡

methods::NTuple{N, Symbol}: The applied methods. The available methods are their configurations (in terms of keyword arguments):

Methods	Configuration(s)	keyword argument(s)	Default value(s)

:DS	Damping strength: [0,1]	dampingStrength::Float64	0.0

:DIIS, :EDIIS, :ADIIS	Subspace size (>1); Coefficient solver(:ADMM-> ADMM solver, :LCM -> Lagrange solver)	DIISsize::Int; solver::Symbol	15; :ADMM

intervals: The stopping (skipping) thresholds for the required methods.

methodConfigs: The additional keywords arguments for each method stored as Tuples of Pairs.

oscillateThreshold: The threshold for oscillating convergence.

≡≡≡ Initialization Method(s) ≡≡≡

SCFconfig(methods::Vector{Symbol}, intervals::Vector{Float64}, 
          configs::Dict{Int, <:Vector{<:Pair}}=Dict(1=>Pair[]);
          oscillateThreshold::Float64=1e-5) -> 
SCFconfig{N}

methods and intervals are the methods to be applied and their stopping (skipping) thresholds respectively; the length of those two Vectors should be the same. configs specifies the additional keyword arguments for each methods by a Pair of which the Int key i is for ith method and the pointed Vector{<:Pair} is the pairs of keyword arguments and their values respectively.

≡≡≡ Example(s) ≡≡≡

julia> SCFconfig([:SD, :ADIIS, :DIIS], [1e-4, 1e-12, 1e-13], Dict(2=>[:solver=>:LCM]) SCFconfig{3}((:SD, :ADIIS, :DIIS), (0.0001, 1.0e-12, 1.0e-13), ((), (:solver => :LCM,), ()), 1.0e-5)

Quiqbox.HFtempVars — Type

HFtempVars{HFtype, N} <: HartreeFockintermediateData

The container to store the intermediate values (only of the same spin configuration) for each iteration during the Hartree-Fock SCF procedure.

≡≡≡ Field(s) ≡≡≡

Cs::Array{Array{T1, 2}, 1} where {Float64<:T1<:Float64}: Coefficient matrices.

Fs::Array{Array{T2, 2}, 1} where {Float64<:T2<:Float64}: Fock matrices

Ds::Array{Array{T3, 2}, 1} where {Float64<:T3<:Float64}: Density matrices corresponding to only spin configuration. For RHF each elements means (unconverged) 0.5*Dᵀ.

Es::Array{Float64, 1}: Part of Hartree-Fock energy corresponding to only spin configuration. For RHF each element means (unconverged) 0.5*E0HF.

shared.Dtots::Array{Array{T, 2}, 1} where {Float64<:T<:Float64}: The total density matrices.

shared.Etots::Array{Float64, 1}: The total Hartree-Fock energy.

NOTE: For UHF, there are 2 HFtempVars being updated during the SCF iterations, and change the field shared.Dtots or shared.Etots of one container will affect the other one's.

Quiqbox.HFfinalVars — Type

HFfinalVars{T, N, Nb} <: HartreeFockFinalValue{T}

The container of the final values after a Hartree-Fock SCF procedure.

≡≡≡ Field(s) ≡≡≡

E0HF::Float64: Hartree-Fock energy of the electronic Hamiltonian.

C::Union{Array{T1, 2}, NTuple{2, Array{T1, 2}}} where {Float64<:T1<:Float64}: Coefficient matrix(s) for one spin configuration.

F::Union{Array{T2, 2}, NTuple{2, Array{T2, 2}}} where {Float64<:T2<:Float64}: Fock matrix(s) for one spin configuration.

D::Union{Array{T3, 2}, NTuple{2, Array{T3, 2}}} where {Float64<:T3<:Float64}: Density matrix(s) for one spin configuration.

Emo::Union{Array{Float64, 1}, NTuple{2, Array{Float64, 1}}}: Energies of molecular orbitals.

occu::Union{Array{Int, 1}, NTuple{2, Array{Int, 1}}}: occupation numbers of molecular orbitals.

temp::Union{HFtempVars{T}, NTuple{2, HFtempVars{T}}} the intermediate values.

isConverged::Bool: Whether the SCF procedure is converged in the end.

Quiqbox.Molecule — Type

Molecule{Nc, Ne, Nb} <:MolecularHartreeFockCoefficient{Nc, Ne}

Container for the information of a molecule.

≡≡≡ Field(s) ≡≡≡

nuc::Tuple{Vararg{String}}: Nuclei of the molecule.

nucCoords::Tuple{Vararg{NTuple{3, Real}}}: The coordinates of the nuclei.

Ne::Int: The number of electrons.

orbital::Tuple{Vararg{MolOrbital}}: Molecular orbitals.

basis::Tuple{Vararg{FloatingGTBasisFuncs}}: The basis set for the molecular orbitals.

E0HF::Float64: Hartree-Fock energy of the electronic Hamiltonian from the basis set.

EnnR::Float64: The nuclear-nuclear repulsion energy.

≡≡≡ Initialization Method(s) ≡≡≡

Molecule(basis::Array{FloatingGTBasisFuncs, 1}, nuc::Array{String, 1}, 
         nucCoords::Array{<:AbstractArray, 1}, Ne::Int, E0HF::Float64, 
         Emos::Array{Float64, 1}, occus::Array{<:Real, 1}, C::Array{Float64, 2}, 
         spins::Array{String, 1}, 
         symms::Array{String, 1}=repeat(["A"], length(occus))) -> 
Molecule{Nc, Ne, Nb}

Emos are the energies of corresponding molecular energies. occus are the occupation numbers of the orbitals. C is the coefficient matrix, which does not need to be a square matrix since the number of rows is the size of the (spatial) basis set whereas the number of the columns represents the number of molecular orbitals. spin specifies the spin functions of the orbitals, entries of which can be set to "Alpha" or "Beta". symms are symmetries of the orbitals where the default entry value is "A" for being antisymmetric.

Quiqbox.MolOrbital — Type

MolOrbital{N} <: AbstractMolOrbital

Struct of molecular orbitals.

≡≡≡ Field(s) ≡≡≡

symmetry::String: The symmetry of the orbital. The default value is "A" for being antisymmetric.

energy::Float64: Molecular energy.

spin::String: Spin function of the orbital. Available values: "Alpha", "Beta".

occupancy::Real: Occupation number.

orbitalCoeffs::NTuple{N, Float64}: coefficients of the basis functions to form the molecular orbital.

≡≡≡ Initialization Method(s) ≡≡≡

MolOrbital(energy::Float64, occupancy::Real, orbitalCoeffs::Array{Float64, 1}, 
           spin::String="Alpha", symmetry::String="A") -> MolOrbital{N}