Molden

Quiqbox supports outputting molecular (in Molecule) information to Molden file format.

Quiqbox.Molden.makeMoldenFile — Function

makeMoldenFile(mol::Molecule; roundDigits::Int=15, 
               recordUMO::Bool=false, fileName::String = "MO") -> String

Write the information of input Molecule into a Molden file. recordUMO determines whether to include the unoccupied molecular orbitals. fileName specifies the name of the file, which is also the returned value. If roundDigits < 0, there won't be rounding for recorded data.

A concrete example of the above function can be found here.