Tool Functions

Quiqbox.checkFname — Function

checkFname(Fname::String; showWarning::Bool=true) -> String

Check if there is a file with the same name in the current directory. If so, will add an "_N" at the end of the file name String. showWarning determines whether prints out the WARNING info when there is a file with the same name.

Quiqbox.hasEqual — Function

hasEqual(obj1, obj2, obj3...; ignoreFunction=false, ignoreContainerType=false) -> Bool

Compare if two objects are the equal.

If ignoreFunction = true then the function will pop up a warning message when a field is a function.

If ignoreContainerType = true then the function will ignore the type difference of the (outermost) container as long as the inside fields are equal.

This function is an instantiation of hasBoolRelation.

≡≡≡ Example(s) ≡≡≡

julia> begin
           struct S
               a::Int
               b::Float64
           end
           a = S(1, 1.0)
           b = S(1, 1.0)
           c = S(1, 1.0)
           d = S(1, 1.1)

           @show hasEqual(a, b, c)
           @show hasEqual(a, b, c, d)
       end
hasEqual(a, b, c) = true
hasEqual(a, b, c, d) = false
false

Quiqbox.hasIdentical — Function

hasIdentical(obj1, obj2, obj3...; 
             ignoreFunction=false, ignoreContainerType=false) -> Bool

Compare if two objects are the Identical. An instantiation of hasBoolRelation.

If ignoreFunction = true then the function will pop up a warning message when a field is a function.

If ignoreContainerType = true then the function will ignore the type difference of the (outermost) container as long as the inside fields are identical.

This function is an instantiation of hasBoolRelation.

≡≡≡ Example(s) ≡≡≡

julia> begin
           struct S
               a::Int
               b::Array{Float64, 1}
           end
            
           a = S(1, [1.0, 1.1])
           b = a
           c = b
           d = S(1, [1.0, 1.1])

           @show hasIdentical(a, b, c)
           @show hasIdentical(a, b, c, d)
       end
hasIdentical(a, b, c) = true
hasIdentical(a, b, c, d) = false
false

Quiqbox.flatten — Function

flatten(a::Tuple) -> Tuple

flatten(a::Array) -> Array

Flatten a::Union{Array, Tuple} that contains Arrays and/or Tuples. Only operate on the outermost layer.

≡≡≡ Example(s) ≡≡≡

julia> flatten((:one, 2, [3, 4.0], ([5], "six"), "7"))
(:one, 2, 3.0, 4.0, [5], "six", "7")

julia> flatten([:one, 2, [3, 4.0], ([5], "six"), "7"])
7-element Vector{Any}:
  :one
 2
 3.0
 4.0
  [5]
  "six"
  "7"

Quiqbox.markUnique — Function

markUnique(arr::AbstractArray, args...; compareFunction::Function = hasEqual, kws...)

Return a markingList using Int number to mark each different elements from (and inside) the input argument(s) and a uniqueList to contain all the unique elements when compareFunction is set to hasEqual (in default).

args and kws are positional arguments and keywords arguments respectively as parameters of the specified compareFunction.

≡≡≡ Example(s) ≡≡≡

julia> markUnique([1, [1, 2],"s", [1, 2]])
([1, 2, 3, 2], Any[1, [1, 2], "s"])

julia> begin 
           struct S
               a::Int
               b::Float64
           end
           
           a = S(1, 2.0)
           b = S(1, 2.0)
           c = S(1, 2.1)
           d = a
           
           markUnique(a,b,c,d)
       end
([1, 1, 2, 1], Any[S(1, 2.0), S(1, 2.1)])

Quiqbox.itself — Function

A function that only returns its argument.