Core Functions

Quiqbox.Exponent — Function

Exponent(e::Union{Real, Array{Float64, 0}}, mapFunction::Function=itself; 
         canDiff::Bool=true, roundDigits::Int=15, dataName::Symbol=:undef) -> 
ParamBox{Float64, :α}

Construct a ParamBox for an exponent coefficient given a value. Keywords mapFunction and canDiff work the same way as in a general constructor of a ParamBox. If roundDigits < 0 or the input e is a 0-d Array, there won't be rounding for input data.

Exponent(pb::ParamBox{Float64}) -> ParamBox{Float64, :α}

Convert a ParamBox to an exponent coefficient parameter.

Quiqbox.Contraction — Function

Contraction(c::Union{Real, Array{Float64, 0}}, mapFunction::Function=itself; 
            canDiff::Bool=true, roundDigits::Int=15, dataName::Symbol=:undef) -> 
ParamBox{Float64, :d}

Construct a ParamBox for an contraction coefficient given a value. Keywords mapFunction and canDiff work the same way as in a general constructor of a ParamBox. If roundDigits < 0 or the input c is a 0-d Array, there won't be rounding for input data.

Contraction(pb::ParamBox{Float64}) -> ParamBox{Float64, :α}

Convert a ParamBox to an contraction coefficient parameter.

Quiqbox.inValOf — Function

inValOf(pb::ParamBox) -> Number

Return the value of stored data (independent variable) of the input ParamBox. Equivalent to pb[].

Quiqbox.outValOf — Function

outValOf(pb::ParamBox) -> Number

Return the value of mapped data (dependent variable) of the input ParamBox. Equivalent to pb().

Quiqbox.inSymOf — Function

inSymOf(pb::ParamBox) -> Symbol

Return the Symbol of the stored data (independent variable) of the input ParamBox.

Quiqbox.outSymOf — Function

outSymOf(pb::ParamBox) -> Symbol

Return the Symbol of the mapped data (dependent variable) of the input ParamBox.

Quiqbox.dataOf — Function

dataOf(pb::ParamBox{T}) where {T<:Number} -> Array{T, 0}

Return the 0-D Array of the data stored in the input ParamBox.

Quiqbox.mapOf — Function

mapOf(pb::ParamBox{<:Number, F}) where {F<:Function} -> F

Return the mapping function of the input ParamBox.

Quiqbox.outValCopy — Function

outValCopy(pb::ParamBox{<:Number, V}) -> ParamBox{<:Number, V, :itself}

Return a new ParamBox of which the stored data is a deep copy of the mapped data from the input ParamBox.

Quiqbox.inVarCopy — Function

inVarCopy(pb::ParamBox) -> ParamBox{<:Number, <:Any, :itself}

Return a new ParamBox of which the stored data is a shallow copy of the stored data from the input ParamBox.

≡≡≡ Example(s) ≡≡≡

julia> pb1 = ParamBox(-2.0, :a, abs)
ParamBox{Float64, :a, :abs}(-2.0)[∂][x_a]

julia> pb2 = inVarCopy(pb1)
ParamBox{Float64, :x_a, :itself}(-2.0)[∂][x_a]

julia> pb1[] == pb2[] == -2.0
true

julia> pb1[] = 1.1
1.1

julia> pb2[]
1.1

Quiqbox.enableDiff! — Function

enableDiff!(pb::ParamBox) -> ParamBox

Mark the input ParamBox as "differentiable" and return the marked ParamBox.

Quiqbox.disableDiff! — Function

disableDiff!(pb::ParamBox) -> ParamBox

Mark the input ParamBox as "non-differentiable" and return the marked ParamBox.

Quiqbox.isDiffParam — Function

isDiffParam(pb::ParamBox) -> Bool

Return the Boolean value of if the input ParamBox is "differentiable".

Quiqbox.toggleDiff! — Function

toggleDiff!(pb::ParamBox) -> Bool

Toggle the differentiability (pb.canDiff[]) of the input ParamBox and return the altered result.

Quiqbox.genBasisFunc — Function

genBasisFunc(args..., kws...) -> BasisFunc

genBasisFunc(args..., kws...) -> BasisFuncs

genBasisFunc(args..., kws...) -> collection

Constructor of BasisFunc and BasisFuncs, but it also returns different kinds of collections of them based on the applied methods.

≡≡≡ Method 1 ≡≡≡

genBasisFunc(coord::AbstractArray, gs::Array{GaussFunc, 1}, 
             ijkOrijks::Union{Array{Int, 1}, Array{Array{Int, 1}, 1}}; 
             normalizeGTO::Bool=false)

ijkOrijks is the Array of the pseudo-quantum number(s) to specify the angular momentum(s). E.g., s is [0,0,0] and p is [[1,0,0], [0,1,0], [0,0,1]].

≡≡≡ Example(s) ≡≡≡

julia> genBasisFunc([0,0,0], GaussFunc(2,1), [0,1,0])
BasisFunc{1, 1}(gauss, subshell, center)[X⁰Y¹Z⁰][0.0, 0.0, 0.0]

≡≡≡ Method 2 ≡≡≡

genBasisFunc(coord::AbstractArray, gs::Union{GaussFunc, Array{GaussFunc, 1}}, 
             subshell::String="S"; 
             ijkFilter::Array{Bool, 1}=fill(true, SubshellDimList[subshell]), 
             normalizeGTO::Bool=false)

≡≡≡ Example(s) ≡≡≡

julia> genBasisFunc([0,0,0], GaussFunc(2,1), "S")
BasisFunc{0, 1}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], GaussFunc(2,1), "P")
BasisFuncs{1, 1, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

≡≡≡ Method 3 ≡≡≡

genBasisFunc(coord::AbstractArray, gExpsANDgCons::NTuple{2, Array{<:Real, 1}}, 
             subshell="S"; kw...)

Instead of directly inputting GaussFunc, one can also input a 2-element Tuple of the exponent(s) and contraction coefficient(s) corresponding to the same GaussFunc(s).

≡≡≡ Example(s) ≡≡≡

julia> genBasisFunc([0,0,0], (2, 1), "P")
BasisFuncs{1, 1, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], ([2, 1.5], [1, 0.5]), "P")
BasisFuncs{1, 2, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

≡≡≡ Method 4 ≡≡≡

genBasisFunc(center, BSKeyANDnuc::Array{Tuple{String, String}, 1})

If the user wants to construct existed atomic basis set(s), they can use the (Array of) (BS_name, Atom_name) as the second input. If the atom is omitted, then basis set for H is used.

≡≡≡ Example(s) ≡≡≡

julia> genBasisFunc([0,0,0], ("STO-3G", "Li"))
3-element Vector{Quiqbox.FloatingGTBasisFuncs}:
 BasisFunc{0, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
 BasisFunc{0, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
 BasisFuncs{1, 3, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], "STO-3G")
1-element Vector{Quiqbox.FloatingGTBasisFuncs}:
 BasisFunc{0, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], ["STO-2G", "STO-3G"])
2-element Vector{Quiqbox.FloatingGTBasisFuncs}:
 BasisFunc{0, 2}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
 BasisFunc{0, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], [("STO-2G", "He"), ("STO-3G", "O")])
4-element Vector{Quiqbox.FloatingGTBasisFuncs}:
 BasisFunc{0, 2}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
 BasisFunc{0, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
 BasisFunc{0, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
 BasisFuncs{1, 3, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

Quiqbox.sortBasisFuncs — Function

sortBasisFuncs(bs::Array{<:FloatingGTBasisFuncs, 1}; groupCenters::Bool=false) ->
Array

Sort basis functions. If groupCenters = true, Then the function will return an Array{<:Array{<:FloatingGTBasisFuncs, 1}, 1} in which the arrays are grouped basis functions with same center coordinates.

Quiqbox.centerOf — Function

centerOf(bf::FloatingGTBasisFuncs) -> Array{<:Real, 1}

Return the center coordinate of the input FloatingGTBasisFuncs.

Quiqbox.add — Method

add(b::CompositeGTBasisFuncs{<:Any, 1}) -> CompositeGTBasisFuncs{<:Any, 1}

add(b1::CompositeGTBasisFuncs{<:Any, 1}, b2::CompositeGTBasisFuncs{<:Any, 1}) ->
CompositeGTBasisFuncs{<:Any, 1}

Addition between CompositeGTBasisFuncs{<:Any, 1} such as BasisFunc and Quiqbox.BasisFuncMix. It can be called using + syntax.

≡≡≡ Example(s) ≡≡≡

julia> bf1 = genBasisFunc([1,1,1], (2,1))
BasisFunc{0, 1}(gauss, subshell, center)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

julia> bf2 = genBasisFunc([1,1,1], (2,2))
BasisFunc{0, 1}(gauss, subshell, center)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

julia> bf3 = bf1 + bf2
BasisFunc{0, 1}(gauss, subshell, center)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

julia> bf3.gauss[1].con[]
3.0

Quiqbox.mul — Method

mul(gf::GaussFunc, coeff::Real) -> GaussFunc

mul(coeff::Real, gf::GaussFunc) -> GaussFunc

mul(gf1::GaussFunc, gf2::GaussFunc) -> GaussFunc

Multiplication between GaussFuncs or contraction coefficient multiplication between a Real number and a GaussFunc. It can be called using * syntax.

≡≡≡ Example(s) ≡≡≡

julia> gf1 = GaussFunc(3,1)
GaussFunc(xpn=ParamBox{Float64, :α, :itself}(3.0)[∂][α], con=ParamBox{Float64, :d, :itself}(1.0)[∂][d])

julia> gf1 * 2
GaussFunc(xpn=ParamBox{Float64, :α, :itself}(3.0)[∂][α], con=ParamBox{Float64, :d, :itself}(2.0)[∂][d])

julia> gf1 * gf1
GaussFunc(xpn=ParamBox{Float64, :α, :itself}(6.0)[∂][α], con=ParamBox{Float64, :d, :itself}(1.0)[∂][d])

julia> gf1 * 2 * gf1
GaussFunc(xpn=ParamBox{Float64, :α, :itself}(6.0)[∂][α], con=ParamBox{Float64, :d, :itself}(2.0)[∂][d])

Quiqbox.mul — Method

mul(sgf1::BasisFunc{𝑙1, 1}, sgf2::BasisFunc{𝑙2, 1}; 
         normalizeGTO::Union{Bool, Missing}=missing)::BasisFunc{𝑙1+𝑙2, 1} where {𝑙1, 𝑙2}

mul(a1::Real, a2::CompositeGTBasisFuncs{<:Any, 1}; 
    normalizeGTO::Union{Bool, Missing}=missing) -> 
CompositeGTBasisFuncs{<:Any, 1}

mul(a1::CompositeGTBasisFuncs{<:Any, 1}, a2::CompositeGTBasisFuncs{<:Any, 1}; 
    normalizeGTO::Union{Bool, Missing}=missing) -> 
CompositeGTBasisFuncs{<:Any, 1}

Multiplication between CompositeGTBasisFuncs{<:Any, 1}s such as BasisFunc and Quiqbox.BasisFuncMix, or contraction coefficient multiplication between a Real number and a CompositeGTBasisFuncs{<:Any, 1}. If normalizeGTO is set to missing (in default), The GaussFunc in the output result will be normalized only if all the input bases have normalizeGTO = true. The function can be called using * syntax.

≡≡≡ Example(s) ≡≡≡

julia> bf1 = genBasisFunc([1,1,1], ([2,1], [0.1, 0.2]))
BasisFunc{0, 2}(gauss, subshell, center)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

julia> bf2 = bf1 * 2
BasisFunc{0, 2}(gauss, subshell, center)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

julia> getindex.(getfield.(bf2.gauss, :con))
(0.2, 0.4)

julia> bf3 = bf1 * bf2
BasisFunc{0, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][1.0, 1.0, 1.0]

Quiqbox.shift — Method

shift(bf::FloatingGTBasisFuncs{𝑙, GN, 1}, didjdk::Array{Int, 1}) where {𝑙, GN} -> 
BasisFunc

Shift (add) the angular momentum (Cartesian representation) given the a vector that specifies the change of each pseudo-quantum number 𝑑i, 𝑑j, 𝑑k.

Quiqbox.decompose — Function

decompose(bf::CompositeGTBasisFuncs; splitGaussFunc::Bool=false) -> 
Array{<:FloatingGTBasisFuncs, 2}

Decompose a FloatingGTBasisFuncs into an Array of BasisFuncs. Each column represents one orbital of the input basis function(s). If splitGaussFunc is true, then each column consists of the BasisFuncs each with only 1 GaussFunc.

Quiqbox.basisSize — Function

basisSize(subshell::Union{String, Array{String, 1}}) -> Tuple

Return the size (number of orbitals) of each subshell.

basisSize(basisFunctions) -> Tuple

Return the numbers of orbitals of the input basis function(s).

Quiqbox.genBasisFuncText — Function

genBasisFuncText(bf::FloatingGTBasisFuncs; norm=1.0, printCenter=true) -> String

Generate a String of the text of the input FloatingGTBasisFuncs. norm is the additional normalization factor. If printCenter is true, the center coordinate will be added on the first line of the String.

genBasisFuncText(bs::Array{<:FloatingGTBasisFuncs, 1}; 
                 norm=1.0, printCenter=true, groupCenters::Bool=true) -> 
String

Generate a String of the text of the input basis set. norm is the additional normalization factor. If printCenter is true, the center coordinate will be added on the first line of the String. groupCenters determines whether the function will group the basis functions with same center together.

Quiqbox.genBFuncsFromText — Function

genBFuncsFromText(content::String; adjustContent::Bool=false, 
                  adjustFunction::F=sciNotReplace, 
                  excludeFirstNlines::Int=0, excludeLastNlines::Int=0, 
                  center::Union{AbstractArray, 
                                Tuple{N, ParamBox}, 
                                Missing}=missing) where {N, F<:Function} -> 
Array{<:FloatingGTBasisFuncs, 1}

Generate the basis set from a String of basis set in Gaussian format or the String output from genBasisFuncText. For the former, adjustContent needs to be set to true. adjustFunction is only applied when adjustContent=true, which in default is a function used to detect and convert the format of the scientific notation in the String.

excludeFirstNlines and excludeLastNlines are used to exclude first or last few lines of the String if intent. center is used to assign a coordinate for all the basis functions from the String; it can be a Vector, a Tuple of the positional ParamBoxs; when it's set to missing, it will try to read the center information from the input string, and leave the center as [NaN, NaN, Nan] if it can't find one for the corresponding BasisFunc. The coordinate information, if included, should be right above the subshell information for the BasisFunc. E.g.:

    """
    X    1.0                 0.0                 0.0
    S    1   1.0
          2.0                 1.0
    """

Quiqbox.assignCenter! — Function

assignCenter!(center::AbstractArray, b::FloatingGTBasisFuncs) -> NTuple{3, ParamBox}

Assign a new coordinate to the center of the input FloatingGTBasisFuncs. Also return the altered center.

Quiqbox.getParams — Function

getParams(pbc::ParamBox, symbol::Union{Symbol, Nothing}=nothing; 
          onlyDifferentiable::Bool=false) -> 
Union{ParamBox, Nothing}

getParams(pbc::StructSpatialBasis, symbol::Union{Symbol, Nothing}=nothing; 
          onlyDifferentiable::Bool=false) -> 
Array{<:ParamBox, 1}

getParams(pbc::Array, symbol::Union{Symbol, Nothing}=nothing; 
          onlyDifferentiable::Bool=false) -> 
Array{<:ParamBox, 1}

Return the parameter(s) stored in the input container. If keyword argument symbol is set to nothing, then return all the different parameters; if it's set to the Symbol of a parameter (e.g. the symbol of ParamBox{T, V} would be V), return only that type of parameters (which might still have different indices). onlyDifferentiable determines whether ignore non-differentiable parameters. If the 1st argument is an Array, the entries must be ParamBox containers.

Quiqbox.copyBasis — Function

copyBasis(b::GaussFunc, copyOutVal::Bool=true) -> GaussFunc

copyBasis(b::CompositeGTBasisFuncs, copyOutVal::Bool=true) -> CompositeGTBasisFuncs

Return a copy of the input basis. If copyOutVal is set to true, then only the value(s) of mapped data will be copied, i.e., outValCopy is used to copy the ParamBoxs, otherwise inVarCopy is used.

≡≡≡ Example(s) ≡≡≡

julia> e = Exponent(3.0, x->x^2)
ParamBox{Float64, :α, :f_α₁}(3.0)[∂][x_α]

julia> c = Contraction(2.0)
ParamBox{Float64, :d, :itself}(2.0)[∂][d]

julia> gf1 = GaussFunc(e, c);

julia> gf2 = copyBasis(gf1)
GaussFunc(xpn=ParamBox{Float64, :α, :itself}(9.0)[∂][α], con=ParamBox{Float64, :d, :itself}(2.0)[∂][d])

julia> gf1.xpn() == gf2.xpn()
true

julia> (gf1.xpn[] |> gf1.xpn.map) == gf2.xpn[]
true

Quiqbox.uniqueParams! — Function

uniqueParams!(bs; filterMapping::Bool=false) -> Array{<:ParamBox, 1}

Mark the parameters (ParamBox) in input bs which can a Vector of GaussFunc or FloatingGTBasisFuncs. The identical parameters will be marked with same index. filterMappingdetermines wether return the ParamBoxs with identical fields except the map field.

Quiqbox.getVar — Function

getVar(pb::ParamBox) -> Symbolics.Num

getVar(container::StructSpatialBasis) -> Array{Symbolics.Num, 1}

Return the independent variable(s) of the input parameter container.

Quiqbox.getVarDict — Function

getVarDict(obj::Union{ParamBox, StructSpatialBasis, Array}; 
           includeMapping::Bool=false) -> 
Dict{Symbolics.Num, <:Number}

Return a Dict that stores the independent variable(s) of the parameter container(s) and its(their) corresponding value(s). If includeMapping = true, then the dictionary will also include the mapping relations between the mapped variables and the independent variables.

Quiqbox.expressionOf — Function

expressionOf(bf::CompositeGTBasisFuncs; splitGaussFunc::Bool=false) -> 
Array{<:Symbolics.Num, 2}

Return the expression(s) of a given CompositeGTBasisFuncs (e.g. BasisFuncMix or FloatingGTBasisFuncs) as a Matrix{<:Symbolics.Num}of which the column(s) corresponds to different orbitals. If splitGaussFunc is true, the column(s) will be expanded vertically such that the entries are GaussFunc inside the corresponding orbital.

expressionOf(gf::GaussFunc) -> Symbolics.Num

Return the expression of a given GaussFunc.

Quiqbox.makeCenter — Function

makeCenter(coord::Array{<:Real, 1}; roundDigits::Int=-1) -> NTuple{3, ParamBox}

Generate a Tuple of coordinate ParamBoxs for a basis function center coordinate given a Vector. If roundDigits < 0 then there won't be rounding for input data.

Quiqbox.gridCoords — Function

gridCoords(gb::GridBox) -> Array{Array{Float64, 1}, 1}

Return the grid-point coordinates in Vectors given the GriBox.

Quiqbox.runHF — Function

runHF(bs::Union{BasisSetData, Array{<:AbstractGTBasisFuncs, 1}}, 
      nuc::Array{String, 1}, 
      nucCoords::Array{<:AbstractArray, 1}, 
      N::Union{NTuple{2, Int}, Int}=getCharge(nuc); 
      initialC::Union{Array{T, 2}, NTuple{2, Array{T, 2}}, Symbol}=:GWH, 
      HFtype::Symbol=:RHF, 
      scfConfig::SCFconfig=defaultSCFconfig, 
      earlyTermination::Bool=true, 
      printInfo::Bool=true, 
      maxSteps::Int=1000) where {Float64<:T<:Float64} -> HFfinalVars

Main function to run Hartree-Fock in Quiqbox.

=== Positional argument(s) ===

bs::Union{BasisSetData, Array{<:AbstractGTBasisFuncs, 1}}: Basis set.

nuc::Array{String, 1}: The element symbols of the nuclei for the Molecule.

nucCoords::Array{<:AbstractArray, 1}: The coordinates of the nuclei.

N::Union{NTuple{2, Int}, Int}: The total number of electrons or the numbers of electrons with different spins respectively.

=== Keyword argument(s) ===

initialC::Union{Array{T, 2}, NTuple{2, Array{T, 2}}, Symbol}: Initial guess of the coefficient matrix(s) C of the molecular orbitals.

HFtype::Symbol: Hartree-Fock type. Available values are :RHF and :UHF.

scfConfig::SCFconfig: SCF iteration configuration.

earlyTermination::Bool: Whether automatically early terminate (skip) a convergence method when its performance becomes unstable or poor.

printInfo::Bool: Whether print out the information of each iteration step.

maxSteps::Int: Maximum allowed iteration steps regardless of whether the SCF converges.

Quiqbox.runHFcore — Function

runHFcore(N::Union{NTuple{2, Int}, Int}, 
          Hcore::Array{T1, 2}, 
          HeeI::Array{T2, 4}, 
          S::Array{T3, 2}, 
          X::Array{T4, 2}=getX(S), 
          C::Union{Array{T, 2}, NTuple{2, Array{T, 2}}, Symbol}=guessC(S, Hcore; X);
          HFtype::Symbol=:RHF,  
          scfConfig::SCFconfig{L}, 
          earlyTermination::Bool=true, 
          printInfo::Bool=true, 
          maxSteps::Int=1000) where {Float64<:T1<:Float64, 
          Float64<:T2<:Float64, 
          Float64<:T3<:Float64, 
          Float64<:T4<:Float64, 
          Float64<:T5<:Float64, L}

The core function of runHF.

=== Positional argument(s) ===

N::Union{NTuple{2, Int}, Int}: The total number of electrons or the numbers of electrons with different spins respectively.

Hcore::Array{T1, 2}: Core Hamiltonian of electronic Hamiltonian.

HeeI::Array{T2, 4}: The electron-electron interaction Hamiltonian which includes both the Coulomb interactions and the Exchange Correlations.

S::Array{T3, 2}: Overlap matrix of the corresponding basis set.

X::Array{T4, 2}: Orthogonal transformation matrix of S. Default value is S^(-0.5).

C::Union{Array{T, 2}, NTuple{2, Array{T, 2}}, Symbol}: Initial guess of the coefficient matrix(s) C of the molecular orbitals.

=== Keyword argument(s) ===

HFtype::Symbol: Hartree-Fock type. Available values are :RHF and :UHF.

scfConfig::SCFconfig: SCF iteration configuration.

earlyTermination::Bool: Whether automatically early terminate (skip) a convergence method when its performance becomes unstable or poor.

printInfo::Bool: Whether print out the information of each iteration step.

maxSteps::Int: Maximum allowed iteration steps regardless of whether the SCF converges.

Quiqbox.getMolOrbitals — Function

getMolOrbitals(ens::Array{Float64, 1}, occus::Array{<:Real, 1}, C::Matrix{Float64}, 
               spins::Array{String, 1}, 
               symms::Array{String, 1}=repeat(["A"], length(occus))) -> 
Tuple{Vararg{getMolOrbitals}}

A function that returns the molecular orbitals.

Quiqbox.nnRepulsions — Function

nnRepulsions(nuc::Array{String, 1}, nucCoords::Array{<:AbstractArray, 1}) -> Float64

Calculate the nuclear-nuclear repulsion energy given the nuclei and their coordinates of a molecule.

Quiqbox.optimizeParams! — Function

optimizeParams!(bs::Array{<:FloatingGTBasisFuncs, 1}, pbs::Array{<:ParamBox, 1},
                nuc::Array{String, 1}, nucCoords::Array{<:AbstractArray, 1}, 
                Ne::Union{NTuple{2, Int}, Int}=getCharge(nuc);
                Etarget::Float64=NaN, threshold::Float64=1e-4, maxSteps::Int=2000, 
                printInfo::Bool=true, GDmethod::F1=gradDescent!, HFtype::Symbol=:RHF, 
                ECmethod::F2=Quiqbox.defaultECmethod) where 
               {F1<:Function, F2<:Function} -> 
Es::Array{Float64, 1}, pars::Array{Float64, 2}, grads::Array{Float64, 2}

The main function to optimize the parameters of a given basis set.

=== Positional argument(s) ===

bs::Array{<:FloatingGTBasisFuncs, 1}: Basis set.

pbs::Array{<:ParamBox, 1}: The parameters to be optimized that are extracted from the basis set.

nuc::Array{String, 1}: The nuclei of the molecule.

nucCoords::Array{<:AbstractArray, 1}: The nuclei coordinates.

Ne::Union{NTuple{2, Int}, Int}: The total number of electrons or the numbers of electrons with different spins respectively.

=== Keyword argument(s) ===

Etarget::Float64: The target Hartree-Hock energy intent to achieve.

threshold::Float64: The threshold for the convergence when evaluating difference between the latest two energies.

maxSteps::Int: Maximum allowed iteration steps regardless of whether the optimization iteration converges.

printInfo::Bool: Whether print out the information of each iteration step.

GDmethod::F1: Applied gradient descent Function.

HFtype::Symbol: Hartree-Fock type. Available values are :RHF and :UHF.

ECmethod::F2: The Function used to update Hartree-Fock energy and coefficient matrix(s) during the optimization iterations.

Quiqbox.updateParams! — Function

updateParams!(pbs::Array{<:ParamBox, 1}, grads::Array{<:Real, 1}; 
              method::F=gradDescent!) where {F<:Function} -> Array{<:ParamBox, 1}

Given a Vector of parameters::ParamBox and its gradients with respect to each parameter, update the ParamBoxs and return the updated values.

Quiqbox.gradDescent! — Function

gradDescent!(pars::Vector{<:Real}, grads::Vector{<:Real}, η=0.001) -> 
pars::Vector{<:Real}

Default gradient descent method in used in Quiqbox.

Quiqbox.overlap — Function

overlap(fb1::AbstractGTBasisFuncs, fb2::AbstractGTBasisFuncs) -> 
Array{Float64, 2}

Return the orbital overlap matrix (an N×N Matrix where N is the number of spatial orbitals) given 2 basis functions.

Quiqbox.overlaps — Function

overlaps(BSet::Array{<:AbstractGTBasisFuncs, 1}) -> Array{Float64, 2}

Return the orbital overlap matrix (an N×N Matrix where N is the number of spatial orbitals) given a basis set in the form of an Array.

Quiqbox.nucAttraction — Function

nucAttraction(fb1::AbstractGTBasisFuncs, fb2::AbstractGTBasisFuncs, 
              nuc::Array{String, 1}, nucCoords::Array{<:AbstractArray, 1}) -> 
Array{Float64, 2}

Return the nuclear attraction matrix (an N×N Matrix where N is the number of spatial orbitals) given 2 basis functions, and the nuclei with their coordinates (in atomic unit).

Quiqbox.nucAttractions — Function

nucAttractions(BSet::Array{<:AbstractGTBasisFuncs, 1}, nuc::Array{String, 1}, 
               nucCoords::Array{<:AbstractArray, 1}) -> 
Array{Float64, 2}

Return the nuclear attraction matrix (an N×N Matrix where N is the number of spatial orbitals) given a basis set in the form of an Array, and the nuclei with their coordinates (in atomic unit).

Quiqbox.elecKinetic — Function

elecKinetic(fb1::AbstractGTBasisFuncs, fb2::AbstractGTBasisFuncs) -> 
Array{Float64, 2}

Return the electron kinetic energy matrix (an N×N Matrix where N is the number of spatial orbitals) given 2 basis functions.

Quiqbox.elecKinetics — Function

elecKinetics(BSet::Array{<:AbstractGTBasisFuncs, 1}) -> Array{Float64, 2}

Return the electron kinetic energy matrix (an N×N Matrix where N is the number of spatial orbitals) given a basis set in the form of an Array.

Quiqbox.coreHij — Function

coreHij(fb1::AbstractGTBasisFuncs, fb2::AbstractGTBasisFuncs) -> 
Array{Float64, 2}

Return a matrix element or block of the core Hamiltonian (an N×N Matrix where N is the number of spatial orbitals) given 2 basis functions.

Quiqbox.coreH — Function

coreH(BSet::Array{<:AbstractGTBasisFuncs, 1}) -> Array{Float64, 2}

Return the core Hamiltonian matrix (an N×N Matrix where N is the number of spatial orbitals) given a basis set in the form of an Array.

Quiqbox.eeInteraction — Function

eeInteraction(bf1::AbstractGTBasisFuncs, 
              bf2::AbstractGTBasisFuncs, 
              bf3::AbstractGTBasisFuncs, 
              bf4::AbstractGTBasisFuncs) -> 
Array{Float64, 4}

Return the electron-electron interaction tensor (an N×N×N×N Tensor where N is the number of spatial orbitals) given 4 basis functions.

Quiqbox.eeInteractions — Function

eeInteractions(BSet::Array{<:AbstractGTBasisFuncs, 1}) -> Array{Float64, 4}

Return the electron-electron interaction tensor (an N×N×N×N Tensor where N is the number of spatial orbitals) given a basis set in the form of an Array.

Quiqbox.eeInteractionsCore — Function

eeInteractionsCore(BSet::Array{<:AbstractGTBasisFuncs, 1}; 
                   outputUniqueIndices::Bool=false) -> 
Array{Float64, 5}, [Array{<:Array{Int, 1}, 1}]

Return the electron-electron interaction tensor (an N×N×N×N×1 Tensor where N is the number of spatial orbitals) given a basis set in the form of an Array.

If outputUniqueIndices=true, additionally return the indices for all the unique integrals.

Quiqbox.oneBodyBFTensor — Function

oneBodyBFTensor(libcinFunc::Symbol, b1::AbstractGTBasisFuncs, 
                b2::AbstractGTBasisFuncs, nuclei::Array{String, 1}=String[], 
                nucleiCoords::Array{<:AbstractArray, 1}=Array[]; 
                isGradient::Bool=false) -> 
Array{Float64, 3}

Core function for one-electron integrals.

libcinFunc::Symbol specifies the backend libcint function name, e.g. "int1e_nuc_cart" should be converted to :int1e_nuc_cart as the input argument. If the integral does not need the information of nuclei and their coordinates, those 2 arguments can be omitted. If the integral is a spacial gradient, isGradient should be set to true.

Quiqbox.twoBodyBFTensor — Function

twoBodyBFTensor(libcinFunc::Symbol, 
                b1::AbstractGTBasisFuncs, b2::AbstractGTBasisFuncs, 
                b3::AbstractGTBasisFuncs, b4::AbstractGTBasisFuncs; 
                isGradient::Bool=false) -> 
Array{Float64, 5}

Core function for one-electron integrals.

libcinFunc::Symbol specifies the backend libcint function name, e.g. "cint2e_cart" should be converted to :cint2e_cart as the input argument.