Core Functions

Quiqbox.genBasisFunc — Function

genBasisFunc(args..., kws...) -> BasisFunc
genBasisFunc(args..., kws...) -> BasisFuncs
genBasisFunc(args..., kws...) -> collection

Constructor of BasisFunc and BasisFuncs, but it also returns different kinds of collections of them based on the applied methods.

≡≡≡ Method 1 ≡≡≡

genBasisFunc(coord::AbstractArray, gs::Array{<:GaussFunc, 1}, 
             ijkOrijks::Union{Array{Int, 1}, Array{Array{Int, 1}, 1}}; 
             normalizeGTO::Bool=false)

ijkOrijks is the Array of the pseudo-quantum number(s) to specify the angular momentum(s). E.g., s is [0,0,0] and p is [[1,0,0], [0,1,0], [0,0,1]].

≡≡≡ Example(s) ≡≡≡

julia> genBasisFunc([0,0,0], GaussFunc(2,1), [0,1,0])
BasisFunc{:P, 1}(gauss, subshell, center)[X⁰Y¹Z⁰][0.0, 0.0, 0.0]

≡≡≡ Method 2 ≡≡≡

genBasisFunc(coord::AbstractArray, gs::Array{<:GaussFunc, 1}, subshell::String="S"; 
             ijkFilter::Array{Bool, 1}=fill(true, SubshellDimList[subshell]), 
             normalizeGTO::Bool=false)

≡≡≡ Example(s) ≡≡≡

julia> genBasisFunc([0,0,0], GaussFunc(2,1), "S")
BasisFunc{:S, 1}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], GaussFunc(2,1), "P")
BasisFuncs{:P, 1, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

≡≡≡ Method 3 ≡≡≡

genBasisFunc(coord::AbstractArray, gExpsANDgCons::NTuple{2, Array{<:Real, 1}}, 
             subshell="S"; kw...)

Instead of directly inputting GaussFunc, one can also input a 2-element Tuple of the exponent(s) and contraction coefficient(s) corresponding to the same GaussFunc(s).

≡≡≡ Example(s) ≡≡≡

julia> genBasisFunc([0,0,0], (2, 1), "P")
BasisFuncs{:P, 1, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], ([2, 1.5], [1, 0.5]), "P")
BasisFuncs{:P, 2, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

≡≡≡ Method 4 ≡≡≡

genBasisFunc(center, BSKeyANDnuc::Array{Tuple{String, String}, 1})

If the user wants to construct existed atomic basis set(s), they can use the (Array of) (BS_name, Atom_name) as the second input. If the atom is omitted, then basis set for H is used.

≡≡≡ Example(s) ≡≡≡

julia> genBasisFunc([0,0,0], ("STO-3G", "Li"))
3-element Vector{Quiqbox.FloatingGTBasisFuncs}:
BasisFunc{:S, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
BasisFunc{:S, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
BasisFuncs{:P, 3, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], "STO-3G")
1-element Vector{Quiqbox.FloatingGTBasisFuncs}:
BasisFunc{:S, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], ["STO-2G", "STO-3G"])
2-element Vector{Quiqbox.FloatingGTBasisFuncs}:
BasisFunc{:S, 2}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
BasisFunc{:S, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]

julia> genBasisFunc([0,0,0], [("STO-2G", "He"), ("STO-3G", "O")])
4-element Vector{Quiqbox.FloatingGTBasisFuncs}:
BasisFunc{:S, 2}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
BasisFunc{:S, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
BasisFunc{:S, 3}(gauss, subshell, center)[X⁰Y⁰Z⁰][0.0, 0.0, 0.0]
BasisFuncs{:P, 3, 3}(gauss, subshell, center)[3/3][0.0, 0.0, 0.0]

Quiqbox.centerOf — Function

centerOf(bf::FloatingGTBasisFuncs) -> Array{<:Real, 1}

Return the center coordinate of the input FloatingGTBasisFuncs.

GTBasis(basis::Vector{<:Quiqbox.AbstractGTBasisFuncs})

Quiqbox.decompose — Function

decompose(bf::CompositeGTBasisFuncs; splitGaussFunc::Bool=false) -> 
Array{<:FloatingGTBasisFuncs, 2}

Decompose a FloatingGTBasisFuncs into an Array of BasisFuncs. Each column represents one orbital of the input basis function(s). If splitGaussFunc is true, then each column consists of the BasisFuncs each with only 1 GaussFunc.

Quiqbox.basisSize — Function

basisSize(subshell::Union{String, Array{String, 1}}) -> Tuple

Return the size (number of orbitals) of each subshell.

basisSize(basisFunctions) -> Tuple

Return the numbers of orbitals of the input basis function(s).

Quiqbox.genBasisFuncText — Function

genBasisFuncText(bf::FloatingGTBasisFuncs; norm=1.0, printCenter=true) -> String

Generate a String of the text of the input FloatingGTBasisFuncs. norm is the additional normalization factor. If printCenter is true, the center coordinate will be added on the first line of the String.

genBasisFuncText(bs::Array{<:FloatingGTBasisFuncs, 1}; 
                 norm=1.0, printCenter=true, groupCenters::Bool=true) -> 
String

Generate a String of the text of the input basis set. norm is the additional normalization factor. If printCenter is true, the center coordinate will be added on the first line of the String. groupCenters determines whether the function will group the basis functions with same center together.

Quiqbox.genBFuncsFromText — Function

genBFuncsFromText(content::String; adjustContent::Bool=false, 
                  adjustFunction::F=sciNotReplace, 
                  excludeFirstNlines::Int=0, excludeLastNlines::Int=0, 
                  center::Union{AbstractArray, 
                                Tuple{N, ParamBox}, 
                                Missing}=missing) where {N, F<:Function} -> 
Array{<:FloatingGTBasisFuncs, 1}

Generate the basis set from a String of basis set in Gaussian format or the String output from genBasisFuncText. For the former, adjustContent needs to be set to true. adjustFunction is only applied when adjustContent=true, which in default is a function used to detect and convert the format of the scientific notation in the String.

excludeFirstNlines and excludeLastNlines are used to exclude first or last few lines of the String if intent. genBFuncsFromText can't directly read center coordinate information from the String even if it's included, so argument center is used to assign a coordinate for all the basis functions from the String; it can be a Vector, a Tuple of the positional ParamBoxs, or simply (in default) set to missing for later assignment.

Quiqbox.assignCenter! — Function

assignCenter!(center::AbstractArray, b::FloatingGTBasisFuncs) -> NTuple{3, ParamBox}

Assign a new coordinate to the center of the input FloatingGTBasisFuncs. Also return the altered center.

Quiqbox.getParams — Function

getParams(pbc::Union{ParamBox, GaussFunc, FloatingGTBasisFuncs}, 
          symbol::Union{Symbol, Nothing}=nothing; onlyDifferentiable::Bool=false) -> 
Union{ParamBox, Array{<:ParamBox, 1}}

Return the parameter(s)::ParamBox stored in the input container. If keyword argument symbol is nothing, then return all the different parameters; if it's set to the Symbol type of a parameter (e.g. the symbol of ParamBox{T} would be T), the function will only search for that type of parameters (which might have different indices). onlyDifferentiable determines whether ignore non-differentiable parameters.

getParams(cs::Array, symbol::Union{Symbol, Nothing}=nothing; 
          onlyDifferentiable::Bool=false) -> 
Array{<:ParamBox, 1}

Method of getParams when the 1st argument is an Array of ParamBox, GaussFunc, FloatingGTBasisFuncs or any of them.

dataCopy

Quiqbox.uniqueParams! — Function

uniqueParams!(bs; ignoreContainer::Bool=false, filter::Bool=true, 
              filterMapping::Bool=false) -> 
Array{<:ParamBox, 1}

Mark the parameters (ParamBox) in input bs which can a Vector of GaussFunc or FloatingGTBasisFuncs. The identical parameters will be marked with same index.

=== Keyword argument(s) ===

ignoreContainer: If set to true, then only the field data of the ParamBoxs will be compared to determine whether each ParamBox are unique.

filter: Determine whether filter out the identical ParamBoxs and only return the unique ones.

filterMapping: Determine wether return the ParamBoxs with identical fields except the map field. When filter=false, this argument is automatically overwritten to be false.

Quiqbox.getVar — Function

getVar(pb::ParamBox; includeMapping::Bool=false) -> 
Array{<:Pair{Symbolics.Num, <:Number}, 1}

Return a 1-element Vector of Pair to show the Symbol::Symbolics.Num of the stored variable and the corresponding values. includeMapping determines whether mappings from the variable to the dependent variable if there is one.

Quiqbox.getVars — Function

getVars(obj::Union{GaussFunc, BasisFunc}; includeMapping::Bool=false) -> 
Array{<:Pair, 1}

getVars(collection::Array{<:Union{GaussFunc, BasisFunc, ParamBox}, 1}; 
        includeMapping::Bool=false) -> 
Array{<:Pair, 1}

Return a Vector of Pair to of the mapping relations between the variables stored in the ParamBoxs and the corresponding values. includeMapping determines whether mappings from the variable(s) to the dependent variable(s) if exists.

Quiqbox.expressionOf — Function

expressionOf(bf::CompositeGTBasisFuncs; splitGaussFunc::Bool=false) -> Array{<:Symbolics.Num, 2}

Return the expression(s) of a given CompositeGTBasisFuncs (e.g. BasisFuncMix or FloatingGTBasisFuncs) as a Matrix{<:Symbolics.Num}of which the column(s) corresponds to different orbitals. If splitGaussFunc is true, the column(s) will be expanded such that its entries are GaussFunc inside the corresponding orbital.

expressionOf(gf::GaussFunc) -> Symbolics.Num

Return the expression of a given GaussFunc.

Quiqbox.GridBox — Type

GridBox(nGridPerEdge::Int, spacing::Real=10, 
        centerCoord::Array{<:Real, 1}=[0.0,0.0,0.0]; 
        canDiff::Bool=true, index::Int=0) -> GridBox

Method of generating a cubic GridBox. nGridPerEdge specifies the number of grids (number of grid points - 1) along each dimension.spacing specifies the length between adjacent grid points. centerCoord specifies the geometry center coordinate of the box. canDiff determines whether the ParamBox should be marked as differentiable. index defines the index number for the actual parameter: spacing L, with the default value 0 it would be L₀.

Quiqbox.makeCenter — Function

makeCenter(coord::Array{<:Real, 1}) -> NTuple{3, ParamBox}

Generate a Tuple of coordinate ParamBoxs for a basis function center coordinate given a Vector.

Quiqbox.gridCoords — Method

gridCoords(gb::GridBox) -> Array{Array{Float64, 1}, 1}

Return the grid-point coordinates in Vectors given the GriBox.

Quiqbox.runHF — Function

runHF(bs::Union{BasisSetData, Array{<:AbstractGTBasisFuncs, 1}}, 
      nuc::Array{String, 1}, 
      nucCoords::Array{<:AbstractArray, 1}, 
      N::Union{NTuple{2, Int}, Int}=getCharge(nuc); 
      initialC::Union{Array{T, 2}, NTuple{2, Array{T, 2}}, Symbol}=:GWH, 
      HFtype::Symbol=:RHF, 
      scfConfig::SCFconfig=defaultSCFconfig, 
      earlyTermination::Bool=true, 
      printInfo::Bool=true, 
      maxSteps::Int=1000) where {Float64<:T<:Float64} -> HFfinalVars

Main function to run Hartree-Fock in Quiqbox.

=== Positional argument(s) ===

bs::Union{BasisSetData, Array{<:AbstractGTBasisFuncs, 1}}: Basis set.

nuc::Array{String, 1}: The element symbols of the nuclei for the Molecule.

nucCoords::Array{<:AbstractArray, 1}: The coordinates of the nuclei.

N::Union{NTuple{2, Int}, Int}: The total number of electrons or the numbers of electrons with different spins respectively.

=== Keyword argument(s) ===

initialC::Union{Array{T, 2}, NTuple{2, Array{T, 2}}, Symbol}: Initial guess of the coefficient matrix(s) C of the molecular orbitals.

HFtype::Symbol: Hartree-Fock type. Available values are :RHF and :UHF.

scfConfig::SCFconfig: SCF iteration configuration.

earlyTermination::Bool: Whether automatically early terminate (skip) a convergence method when its performance becomes unstable or poor.

printInfo::Bool: Whether print out the information of each iteration step.

maxSteps::Int: Maximum allowed iteration steps regardless of whether the SCF converges.

Quiqbox.runHFcore — Function

runHFcore(N::Union{NTuple{2, Int}, Int}, 
          Hcore::Array{T1, 2}, 
          HeeI::Array{T2, 4}, 
          S::Array{T3, 2}, 
          X::Array{T4, 2}=getX(S), 
          C::Union{Array{T, 2}, NTuple{2, Array{T, 2}}, Symbol}=guessC(S, Hcore; X);
          HFtype::Symbol=:RHF,  
          scfConfig::SCFconfig{L}, 
          earlyTermination::Bool=true, 
          printInfo::Bool=true, 
          maxSteps::Int=1000) where {Float64<:T1<:Float64, 
          Float64<:T2<:Float64, 
          Float64<:T3<:Float64, 
          Float64<:T4<:Float64, 
          Float64<:T5<:Float64, L}

The core function of runHF.

=== Positional argument(s) ===

N::Union{NTuple{2, Int}, Int}: The total number of electrons or the numbers of electrons with different spins respectively.

Hcore::Array{T1, 2}: Core Hamiltonian of electronic Hamiltonian.

HeeI::Array{T2, 4}: The electron-electron interaction Hamiltonian which includes both the Coulomb interactions and the Exchange Correlations.

S::Array{T3, 2}: Overlap matrix of the corresponding basis set.

X::Array{T4, 2}: Orthogonal transformation matrix of S. Default value is S^(-0.5).

C::Union{Array{T, 2}, NTuple{2, Array{T, 2}}, Symbol}: Initial guess of the coefficient matrix(s) C of the molecular orbitals.

=== Keyword argument(s) ===

HFtype::Symbol: Hartree-Fock type. Available values are :RHF and :UHF.

scfConfig::SCFconfig: SCF iteration configuration.

earlyTermination::Bool: Whether automatically early terminate (skip) a convergence method when its performance becomes unstable or poor.

printInfo::Bool: Whether print out the information of each iteration step.

maxSteps::Int: Maximum allowed iteration steps regardless of whether the SCF converges.

Quiqbox.Molecule — Method

Molecule(basis::Array{<:FloatingGTBasisFuncs, 1}, nuc::Array{String, 1}, 
         nucCoords::Array{<:AbstractArray, 1}, HFfVars::HFfinalVars) -> Molecule

Construct a Molecule from a basis set, nuclei information, and the result from the corresponding Hartree-Fock SCF procedure, specifically a HFfinalVars struct.

Quiqbox.getMolOrbitals — Function

getMolOrbitals(ens::Array{Float64, 1}, occus::Array{<:Real, 1}, C::Matrix{Float64}, 
               spins::Array{String, 1}, 
               symms::Array{String, 1}=repeat(["A"], length(occus))) -> 
Tuple{Vararg{getMolOrbitals}}

A function that returns the molecular orbitals.

Quiqbox.nnRepulsions — Function

nnRepulsions(nuc::Array{String, 1}, nucCoords::Array{<:AbstractArray, 1}) -> Float64

Calculate the nuclear-nuclear repulsion energy given the nuclei and their coordinates of a molecule.

Quiqbox.optimizeParams! — Function

optimizeParams!(bs::Array{<:FloatingGTBasisFuncs, 1}, pbs::Array{<:ParamBox, 1},
                nuc::Array{String, 1}, nucCoords::Array{<:AbstractArray, 1}, 
                Ne::Union{NTuple{2, Int}, Int}=getCharge(nuc);
                Etarget::Float64=NaN, threshold::Float64=1e-4, maxSteps::Int=2000, 
                printInfo::Bool=true, GDmethod::F1=gradDescent!, HFtype::Symbol=:RHF, 
                ECmethod::F2=Quiqbox.defaultECmethod) where 
               {F1<:Function, F2<:Function} -> 
Es::Array{Float64, 1}, pars::Array{Float64, 2}, grads::Array{Float64, 2}

The main function to optimize the parameters of a given basis set.

=== Positional argument(s) ===

bs::Array{<:FloatingGTBasisFuncs, 1}: Basis set.

pbs::Array{<:ParamBox, 1}: The parameters to be optimized that are extracted from the basis set.

nuc::Array{String, 1}: The nuclei of the molecule.

nucCoords::Array{<:AbstractArray, 1}: The nuclei coordinates.

Ne::Union{NTuple{2, Int}, Int}: The total number of electrons or the numbers of electrons with different spins respectively.

=== Keyword argument(s) ===

Etarget::Float64: The target Hartree-Hock energy intent to achieve.

threshold::Float64: The threshold for the convergence when evaluating difference between the latest two energies.

maxSteps::Int: Maximum allowed iteration steps regardless of whether the optimization iteration converges.

printInfo::Bool: Whether print out the information of each iteration step.

GDmethod::F1: Applied gradient descent Function.

HFtype::Symbol: Hartree-Fock type. Available values are :RHF and :UHF.

ECmethod::F2: The Function used to update Hartree-Fock energy and coefficient matrix(s) during the optimization iterations. === Keyword argument(s) ===

Quiqbox.updateParams! — Function

updateParams!(pbs::Array{<:ParamBox, 1}, grads::Array{<:Real, 1}; 
              method::F=gradDescent!) where {F<:Function} -> Array{<:ParamBox, 1}

Given a Vector of parameters::ParamBox and its gradients with respect to each parameter, update the ParamBoxs and return the updated values.

Quiqbox.gradDescent! — Function

gradDescent!(pars::Vector{<:Real}, grads::Vector{<:Real}, η=0.001) -> 
pars::Vector{<:Real}

Default gradient descent method in used in Quiqbox.

Quiqbox.overlap — Function

overlap(fb1::AbstractGTBasisFuncs, fb2::AbstractGTBasisFuncs) -> 
Array{Float64, 2}

Return the orbital overlap matrix (an N×N Matrix where N is the number of spatial orbitals) given 2 basis functions.

Quiqbox.overlaps — Function

overlaps(BSet::Array{<:AbstractGTBasisFuncs, 1}) -> Array{Float64, 2}

Return the orbital overlap matrix (an N×N Matrix where N is the number of spatial orbitals) given a basis set in the form of an Array.

Quiqbox.nucAttraction — Function

nucAttraction(fb1::AbstractGTBasisFuncs, fb2::AbstractGTBasisFuncs, 
              nuc::Array{String, 1}, nucCoords::Array{<:AbstractArray, 1}) -> 
Array{Float64, 2}

Return the nuclear attraction matrix (an N×N Matrix where N is the number of spatial orbitals) given 2 basis functions, and the nuclei with their coordinates (in atomic unit).

Quiqbox.nucAttractions — Function

nucAttractions(BSet::Array{<:AbstractGTBasisFuncs, 1}, nuc::Array{String, 1}, 
               nucCoords::Array{<:AbstractArray, 1}) -> 
Array{Float64, 2}

Return the nuclear attraction matrix (an N×N Matrix where N is the number of spatial orbitals) given a basis set in the form of an Array, and the nuclei with their coordinates (in atomic unit).

Quiqbox.elecKinetic — Function

elecKinetic(fb1::AbstractGTBasisFuncs, fb2::AbstractGTBasisFuncs) -> 
Array{Float64, 2}

Return the electron kinetic energy matrix (an N×N Matrix where N is the number of spatial orbitals) given 2 basis functions.

Quiqbox.elecKinetics — Function

elecKinetics(BSet::Array{<:AbstractGTBasisFuncs, 1}) -> Array{Float64, 2}

Return the electron kinetic energy matrix (an N×N Matrix where N is the number of spatial orbitals) given a basis set in the form of an Array.

Quiqbox.coreHij — Function

coreHij(fb1::AbstractGTBasisFuncs, fb2::AbstractGTBasisFuncs) -> 
Array{Float64, 2}

Return a matrix element or block of the core Hamiltonian (an N×N Matrix where N is the number of spatial orbitals) given 2 basis functions.

Quiqbox.coreH — Function

coreH(BSet::Array{<:AbstractGTBasisFuncs, 1}) -> Array{Float64, 2}

Return the core Hamiltonian matrix (an N×N Matrix where N is the number of spatial orbitals) given a basis set in the form of an Array.

Quiqbox.eeInteraction — Function

eeInteraction(bf1::AbstractGTBasisFuncs, 
              bf2::AbstractGTBasisFuncs, 
              bf3::AbstractGTBasisFuncs, 
              bf4::AbstractGTBasisFuncs) -> 
Array{Float64, 4}

Return the electron-electron interaction tensor (an N×N×N×N Tensor where N is the number of spatial orbitals) given 4 basis functions.

Quiqbox.eeInteractions — Function

eeInteractions(BSet::Array{<:AbstractGTBasisFuncs, 1}) -> Array{Float64, 4}

Return the electron-electron interaction tensor (an N×N×N×N Tensor where N is the number of spatial orbitals) given a basis set in the form of an Array.

Quiqbox.eeInteractionsCore — Method

eeInteractionsCore(BSet::Array{<:AbstractGTBasisFuncs, 1}; 
                   outputUniqueIndices::Bool=false) -> 
Array{Float64, 5}, [Array{<:Array{Int, 1}, 1}]

Return the electron-electron interaction tensor (an N×N×N×N×1 Tensor where N is the number of spatial orbitals) given a basis set in the form of an Array.

If outputUniqueIndices=true, additionally return the indices for all the unique integrals.

Quiqbox.oneBodyBFTensor — Function

oneBodyBFTensor(libcinFunc::Symbol, b1::AbstractGTBasisFuncs, 
                b2::AbstractGTBasisFuncs, nuclei::Array{String, 1}=String[], 
                nucleiCoords::Array{<:AbstractArray, 1}=Array[]; 
                isGradient::Bool=false) -> 
Array{Float64, 2}

Core function for one-electron integrals.

libcinFunc::Symbol specifies the backend libcint function name, e.g. "int1e_nuc_cart" should be converted to :int1e_nuc_cart as the input argument. If the integral does not need the information of nuclei and their coordinates, those 2 arguments can be omitted. If the integral is a spacial gradient, isGradient should be set to true.

Quiqbox.twoBodyBFTensor — Function

twoBodyBFTensor(libcinFunc::Symbol, 
                b1::AbstractGTBasisFuncs, b2::AbstractGTBasisFuncs, 
                b3::AbstractGTBasisFuncs, b4::AbstractGTBasisFuncs; 
                isGradient::Bool=false) -> 
Array{Float64, 5}

Core function for one-electron integrals.

libcinFunc::Symbol specifies the backend libcint function name, e.g. "cint2e_cart" should be converted to :cint2e_cart as the input argument.