Core Types

Quiqbox.ParamBox — Type

ParamBox{T, V, FL<:FLevel} <: DifferentiableParameter{T, ParamBox}

Parameter container that can enable differentiation.

≡≡≡ Field(s) ≡≡≡

data::Array{T, 0}: The container of the data (i.e. the value of the input variable) stored in the ParamBox that can be accessed by syntax [].

dataName::Symbol: The name of the input variable.

map::Function: The mapping of the data within the same domain (.map(::T)->T). The result (i.e., the value of the output variable) can be accessed by syntax ().

canDiff::Array{Bool, 0}: Indicator of whether the output variable is "marked" as differentiable with respect to the input variable in the differentiation process. In other words, it determines whether the output variable represented by the ParamBox is treated as a dependent variable or an independent variable.

index::Array{<:Union{Int, Nothing}, 0}: Additional index assigned to the ParamBox.

≡≡≡ Initialization Method(s) ≡≡≡

ParamBox(data, dataName=:undef; index=nothing) -> ParamBox{<:Any, dataName, FLevel{0}}

ParamBox(data, name, mapFunction, dataName=:undef; index=nothing, canDiff=true) ->
ParamBox{<:Any, name}

=== Positional argument(s) ===

data::Union{Array{T, 0}, T}: The input value to be stored or the container of it. If the latter is the first argument, then it will directly be assigned to .data.

name::Symbol: Specify the name of the output variable represented by the constructed ParamBox. It's not required when mapFunction is not provided because then the output variable is considered the same as the input variable. It's equal to the type parameter V of the constructed ParamBox.

mapFunction::Function: The mapping of the stored data (mapFunction(::T)->T), which will be assigned to the field .map. After constructing a ParamBox, e.g pb = ParamBox(x, yName, f), pb[] returns the value of x, and pb() returns the value of f(x)::T. When mapFunction is not provided, .map is set to itself that maps the stored data to itself.

dataName::Symbol: The name of the stored data, i.e, the name of the input variable.

=== Keyword argument(s) ===

index::Union{Int, Nothing}: The index of the constructed ParamBox. It's should be left with its default value unless the user plans to utilize the index of a ParamBox for specific application other than differentiation.

canDiff::Bool: Determine whether the output variable is marked as "differentiable".

≡≡≡ Example(s) ≡≡≡

julia> ParamBox(1.0)
ParamBox{Float64, :undef, FLevel{0}}(1.0)[∂][undef]

julia> ParamBox(1.0, :a)
ParamBox{Float64, :a, FLevel{0}}(1.0)[∂][a]

julia> ParamBox(1.0, :a, abs)
ParamBox{Float64, :a, FLevel{1}}(1.0)[∂][x_a]

NOTE 1: The rightmost "[∂][IV]" in the printed info indicates the differentiability and the name (with an assigned index) of the independent variable held by the ParamBox. When the ParamBox is marked as a "differentiable parameter", "[∂]" is bold and green instead of just being grey, and IV is the name of the input variable.

NOTE 2: The output variable of a ParamBox is normally used to differentiate a parameter functional (e.g., the Hartree-Fock energy). However, the derivative with respect to the corresponding input variable can also be computed to when the ParamBox is marked as differentiable.

Quiqbox.BasisFunc — Type

BasisFunc{T, D, 𝑙, GN, PT} <: FloatingGTBasisFuncs{T, D, 𝑙, GN, PT, 1}

A (floating) Gaussian-type basis function with its center assigned to a defined coordinate.

≡≡≡ Field(s) ≡≡≡

center::SpatialPoint{T, D, PT}: The D-dimensional center.

gauss::NTuple{N, GaussFunc{T, <:Any}}: Gaussian functions within the basis function.

l::Tuple{LTuple{D, 𝑙}}: Cartesian representation of the angular momentum. E.g., LTuple{3, 1}((1, 0, 0)) (X¹Y⁰Z⁰) would correspond to an specific angular momentum configuration where the sum of all the components is 𝑙=1.

normalizeGTO::Bool: Whether each GaussFunc inside will be normalized in calculations.

param::NTuple{D+GN*2, ParamBox}： All the tunable parameters stored in the basis function.

≡≡≡ Initialization Method(s) ≡≡≡

BasisFunc(cen::SpatialPoint{T, D, PT}, gs::NTuple{GN, AbstractGaussFunc{T}}, 
          l::Union{Tuple{LTuple{D, 𝑙}}, LTuple{D, 𝑙}}, normalizeGTO::Bool) where 
         {T, D, PT, 𝑙, GN} -> 
BasisFunc{T, D, 𝑙, GN, PT}

BasisFunc(cen::SpatialPoint{T, D, PT}, gs::AbstractGaussFunc{T}, 
          l::Union{Tuple{LTuple{D, 𝑙}}, LTuple{D, 𝑙}}, normalizeGTO::Bool) where 
         {T, D, PT, 𝑙, GN} -> 
BasisFunc{T, D, 𝑙, 1, PT}

Quiqbox.BasisFuncMix — Type

BasisFuncMix{T, D, BN, BFT<:BasisFunc{T, D}} <: CompositeGTBasisFuncs{T, D, BN, 1}

Sum of multiple FloatingGTBasisFuncs{<:Any, <:Any, <:Any, <:Any, <:Any, 1} without any reformulation, treated as one basis function in the calculations.

≡≡≡ Field(s) ≡≡≡

BasisFunc::NTuple{BN, BFT}: Basis functions used to sum up.

param::Tuple{Vararg{ParamBox}}: Contained parameters.

≡≡≡ Initialization Method(s) ≡≡≡

BasisFuncMix(bfs::Union{Tuple{Vararg{T}}, AbstractArray{T}}) where 
            {T<:FloatingGTBasisFuncs{<:Any, <:Any, <:Any, <:Any, <:Any, 1}} -> 
BasisFuncMix

Quiqbox.BasisFuncs — Type

BasisFuncs{T, D, 𝑙, GN, PT, ON} <: FloatingGTBasisFuncs{T, D, 𝑙, GN, PT, ON}

A collection of basis functions with identical parameters except having different orientations within a specified subshell (i.e. same total orbital angular momentum). It has the same fields as BasisFunc. Specifically, for l, its size ON can be no less than 1 and no larger than the size of the corresponding subshell.

Quiqbox.GTBasis — Type

GTBasis{T, D, BN, BFT<:GTBasisFuncs{T, D, 1}} <: BasisSetData{T, D, BFT}

The container of basis set information.

≡≡≡ Field(s) ≡≡≡

basis::NTuple{BN, BFT}: Stored basis set.

S::Matrix{T}: Overlap matrix.

Te::Matrix{T}: Kinetic energy part of the electronic core Hamiltonian.

eeI::Array{T, 4}: Electron-electron interaction.

≡≡≡ Initialization Method(s) ≡≡≡

GTBasis(basis::Union{Tuple{Vararg{GTBasisFuncs{T, D}}}, 
                     AbstractVector{<:GTBasisFuncs{T, D}}}) where {T, D} -> 
GTBasis{T, D}

Construct a GTBasis given a basis set.

Quiqbox.GaussFunc — Type

GaussFunc{T, FL1, FL2} <: AbstractGaussFunc{T}

A contracted primitive Gaussian-type function.

≡≡≡ Field(s) ≡≡≡

xpn::ParamBox{T, :α, FL1}：The exponent coefficient.

con::ParamBox{T, :d, FL2}: The contraction coefficient.

param::Tuple{ParamBox{T, α}, ParamBox{T, d}}: The parameter containers inside a GaussFunc.

≡≡≡ Initialization Method(s) ≡≡≡

GaussFunc(e::Union{T, ParamBox{T}}, d::Union{T, ParamBox{T}}) where 
         {T<:AbstractFloat} -> 
GaussFunc{T}

Quiqbox.SpatialPoint — Type

SpatialPoint{T, D, PT} <: AbstractSpatialPoint{T, D}

A D-dimensional spatial point.

≡≡≡ Field(s) ≡≡≡

param::PT: A Tuple of ParamBoxs as the components of the spatial coordinate.

≡≡≡ Initialization Method(s) ≡≡≡

SpatialPoint(pbs::Union{Tuple{ParamBox{T, :X, FLevel{Lx}}} where Lx, Tuple{ParamBox{T, :X, FLevel{Lx}}, ParamBox{T, :Y, FLevel{Ly}}} where {Lx, Ly}, Tuple{ParamBox{T, :X, FLevel{Lx}}, ParamBox{T, :Y, FLevel{Ly}}, ParamBox{T, :Z, FLevel{Lz}}} where {Lx, Ly, Lz}} where T) -> SpatialPoint

Quiqbox.HFconfig — Type

HFconfig{T1, HFT, F, T2, L} <: ConfigBox{T1, HFconfig, HFT}

The container of Hartree-Fock method configuration.

≡≡≡ Field(s) ≡≡≡

HF::Val{HFT}: Hartree-Fock method type. Available values of HFT are :RHF, :UHF.

C0::InitialC{T1, HFT, F}: Initial guess of the coefficient matrix(s) C of the canonical orbitals. When C0 is a Val{T}, the available values of T are :GWH, :Hcore, :SAD.

SCF::SCFconfig{T2, L}: SCF iteration configuration. For more information please refer to SCFconfig.

maxStep::Int: Maximum iteration steps allowed regardless if the iteration converges.

earlyStop::Bool: Whether automatically terminate (or skip) a convergence method early when its performance becomes unstable or poor.

≡≡≡ Initialization Method(s) ≡≡≡

HFconfig(;HF::Symbol=:RHF, 
          C0::Union{Tuple{AbstractMatrix}, 
                    NTuple{2, AbstractMatrix}, Symbol}=:SAD, 
          SCF::SCFconfig=SCFconfig{Float64, 2}(method=(:ADIIS, :DIIS), interval=(0.005, 1.0e-12), methodConfig, oscillateThreshold), 
          maxStep::Int=100, 
          earlyStop::Bool=true) -> 
HFconfig

≡≡≡ Example(s) ≡≡≡

julia> HFconfig();

julia> HFconfig(HF=:UHF);

Quiqbox.HFfinalVars — Type

HFfinalVars{T, D, HFT, NN, BN, HFTS} <: HartreeFockFinalValue{T, HFT}

The container of the final values after a Hartree-Fock SCF procedure.

≡≡≡ Field(s) ≡≡≡

Ehf::T: Hartree-Fock energy of the electronic Hamiltonian.

Enn::T: The nuclear repulsion energy.

Ns::NTuple{2, Int}: The numbers of two different spins respectively.

nuc::NTuple{NN, String}: The nuclei in the studied system.

nucCoords::NTuple{NN, NTuple{D, T}}: The coordinates of corresponding nuclei.

C::NTuple{HFTS, Matrix{T}}: Coefficient matrix(s) for one spin configuration.

D::NTuple{HFTS, Matrix{T}}: Density matrix(s) for one spin configuration.

F::NTuple{HFTS, Matrix{T}}: Fock matrix(s) for one spin configuration.

Eo::NTuple{HFTS, Vector{T}}: Energies of canonical orbitals.

occu::NTuple{HFTS, NTuple{BN, String}}: Occupations of canonical orbitals.

temp::NTuple{HFTS, HFtempVars{T, HFT}}: the intermediate values.

isConverged::Bool: Whether the SCF procedure is converged in the end.

basis::GTBasis{T, D, BN}: The basis set used for the Hartree-Fock approximation.

Quiqbox.HFtempVars — Type

HFtempVars{T, HFT} <: HartreeFockintermediateData{T}

The container to store the intermediate values (only of the one spin configuration) for each iteration during the Hartree-Fock SCF procedure.

≡≡≡ Field(s) ≡≡≡

N::Int: The number of electrons with the one spin function.

Cs::Vector{Matrix{T}}: Coefficient matrices.

Ds::Vector{Matrix{T}}: Density matrices corresponding to only spin configuration.

Fs::Vector{Matrix{T}}: Fock matrices.

Es::Vector{T}: Part of the Hartree-Fock energy corresponding to one spin configuration.

shared.Dtots::Vector{Matrix{T}}: The total density matrices.

shared.Etots::Vector{T}: The total Hartree-Fock energy.

NOTE: For unrestricted Hartree-Fock, there are 2 HFtempVars being updated during the iterations, and changing the field shared.Dtots or shared.Etots of one HFtempVars will affect the other one's.

Quiqbox.SCFconfig — Type

SCFconfig{T, L} <: ImmutableParameter{T, SCFconfig}

The struct for self-consistent field (SCF) iteration configurations.

≡≡≡ Field(s) ≡≡≡

method::NTuple{L, Symbol}: The applied convergence methods. The available methods and their configurations (in terms of keyword arguments):

Convergence Method(s)	Configuration(s)	Keyword(s)	Range(s)/Option(s)	Default(s)

:DD	damping strength	dampStrength	[0, 1]	0.5

:DIIS, :EDIIS, :ADIIS	subspace size; coefficient solver	DIISsize; solver	1,2...; :LCM, :LBFGS	12; :LBFGS

Convergence Methods

DD: Direct diagonalization of the Fock matrix.

DIIS: Direct inversion in the iterative subspace.

EDIIS: Energy-DIIS.

ADIIS: DIIS based on the augmented Roothaan–Hall (ARH) energy function.

DIIS-type Method Solvers

LCM: Lagrange multiplier solver.

LBFGS: Limited-memory BFGS with box constraints.

interval::NTuple{L, T}: The stopping (or skipping) thresholds for required methods.

methodConfig::NTuple{L, Vector{<:Pair}}: The additional keywords arguments for each method stored as Tuples of Pairs.

oscillateThreshold::T: The threshold for oscillating convergence.

≡≡≡ Initialization Method(s) ≡≡≡

SCFconfig(methods::NTuple{L, Symbol}, intervals::NTuple{L, T}, 
          config::Dict{Int, <:AbstractVector{<:Pair}}=Dict(1=>Pair[]);
          oscillateThreshold::Real=1.0e-6) where {L, T} -> 
SCFconfig{T, L}

methods and intervals are the convergence methods to be applied and their stopping (or skipping) thresholds respectively. config specifies additional keyword argument(s) for each methods by a Pair of which the key i::Int is for ith method and the pointed AbstractVector{<:Pair} is the pairs of keyword arguments and their values respectively.

SCFconfig(;threshold::AbstractFloat=(0.005, 1.0e-12), 
           oscillateThreshold::Real=defultOscThreshold) -> 
SCFconfig{Float64, 2}

threshold will update the stopping threshold of the default SCF configuration used in HFconfig() with a new value. In other words, it updates the stopping threshold of :1.0e-12.

≡≡≡ Example(s) ≡≡≡

julia> SCFconfig((:DD, :ADIIS, :DIIS), (1e-4, 1e-12, 1e-13), Dict(2=>[:solver=>:LCM]));

julia> SCFconfig(threshold=1e-8, oscillateThreshold=1e-5)
SCFconfig{Float64, 2}(method=(:ADIIS, :DIIS), interval=(0.005, 1.0e-8), methodConfig, oscillateThreshold)

Quiqbox.GridBox — Type

GridBox{T, D, NP, GPT<:SpatialPoint{T, D}} <: SpatialStructure{T, D}

A container of multiple D-dimensional grid points.

≡≡≡ Field(s) ≡≡≡

spacing::T: The distance between adjacent grid points, a.k.a the edge length of each grid.

nPoint::Int: Total number of the grid points.

point::NTuple{NP, GPT}: The grid points represented by SpatialPoint.

param::Tuple{Vararg{ParamBox{T}}}: All the parameters in the GridBox.

≡≡≡ Initialization Method(s) ≡≡≡

GridBox(nGrids::NTuple{D, Int}, spacing::Union{T, Array{T, 0}}, 
        center::Union{AbstractVector{T}, NTuple{D, T}}=ntuple(_->T(0),Val(D)); 
        canDiff::Bool=true, index::Int=0) where {T<:AbstractFloat, D} -> 
GridBox{T, D}

Construct a general D-dimensional GridBox.

=== Positional argument(s) ===

nGrids::NTuple{D, Int}: The numbers of grids along each dimension.

spacing::Union{T, Array{T, 0}}: The edge length of each grid.

center::Union{AbstractVector{T}, NTuple{D, T}}: The coordinate of the geometric center of the grid box.

=== Keyword argument(s) ===

canDiff: If all the ParamBoxes stored in the constructed GridBox will be marked as differentiable.

index: The Index that will be assigned to the shared input variable L of all the stored ParamBoxes.

Quiqbox.GridBox — Method

GridBox(nGridPerEdge::Int, spacing::T, 
        center::Union{AbstractVector{T}, NTuple{D, T}}=ntuple(_->T(0), 3); 
        canDiff::Bool=true, index::Int=0) where {T<:AbstractFloat, D} -> 
GridBox{T, D}

The method of generating a cubic GridBox. Aside from the common arguments, nGridPerEdge specifies the number of grids for every dimension. The dimension of the grid box is determined by the dimension of center.

Quiqbox.POconfig — Type

POconfig{T, M, CBT<:ConfigBox, F<:Function} <: ConfigBox{T, POconfig, M}

The mutable container of parameter optimization configurations.

≡≡≡ Field(s) ≡≡≡

method::Val{M}: The method to calculate objective function (e.g., HF energy) for optimization. Available values of M from Quiqbox are :HF,.

config::CBT: The configuration for the selected method. E.g., for :HF it's HFconfig.

target::T: The value of target function aimed to achieve. If it's set to NaN, the convergence will be solely based on threshold between the latest updated function values rather then the latest value and target.

threshold::T: The error threshold (i.e. the absolute value of the function loss with respect to target) for the convergence; when set to NaN, there will be no convergence detection.

maxStep::Int: Maximum iteration steps allowed regardless if the iteration converges.

GD::F: Applied gradient descent Function. Default method is gradDescent!.

≡≡≡ Initialization Method(s) ≡≡≡

POconfig(;method=Val{:HF}(), 
          config=HFconfig(), 
          target=NaN, 
          threshold=5.0e-7, 
          maxStep=500, 
          GD=gradDescent!) -> 
POconfig

≡≡≡ Example(s) ≡≡≡

julia> POconfig();

julia> POconfig(maxStep=100);

Quiqbox.CanOrbital — Type

CanOrbital{T, D, NN} <: AbstractSpinOrbital{T, D}

The spatial part (orbital) of a canonical spin-orbital (the set of which diagonalizes the Fock matrix of a Hartree-Fock state) with its occupation information. This means the maximal occupation number for the mode corresponding to the orbital (namely a canonical orbital) equals 2. Please refer to getCanOrbitals for the construction of a CanOrbital.

≡≡≡ Field(s) ≡≡≡

energy::T: The eigen energy corresponding to the orbital.

index::Int: The index of the orbital within the same spin configuration.

nuc::NTuple{NN, String}: The nuclei in the studied system.

nucCoords::NTuple{NN, NTuple{D, T}}: The coordinates of corresponding nuclei.

occu::NTuple{2, Array{Bool, 0}}: The occupations of two spin configurations.

orbital::GTBasisFuncs{T, D, 1}: The spatial orbital part.

Quiqbox.MatterByHF — Type

MatterByHF{T, D, NN, N, BN, HFTS} <:MatterData{T, D, N}

Container of the electronic structure information of a quantum system.

≡≡≡ Field(s) ≡≡≡

Ehf::T: Hartree-Fock energy of the electronic Hamiltonian.

nuc::NTuple{NN, String}: The nuclei in the studied system.

nucCoord::NTuple{NN, NTuple{D, T}}: The coordinates of corresponding nuclei.

Enn::T: The nuclear repulsion energy.

Ns::NTuple{2, Int}: The numbers of two different spins respectively.

occuOrbital::NTuple{HFTS, Tuple{Vararg{CanOrbital{T, D, NN}}}}: The occupied canonical orbitals.

unocOrbital::NTuple{HFTS, Tuple{Vararg{CanOrbital{T, D, NN}}}} The unoccupied canonical orbitals.

occuC::NTuple{HFTS, Matrix{T}}: Coefficient matrix(s) of occupied canonical orbitals.

unocC::NTuple{HFTS, Matrix{T}}: Coefficient matrix(s) of unoccupied canonical orbitals.

coreHsameSpin::NTuple{HFTS, Matrix{T}}: Core Hamiltonian(s) (one-body integrals) of the canonical orbitals with same spin configuration(s).

eeIsameSpin::NTuple{HFTS, Array{T, 4}}: electron-electron interactions (two-body integrals) of the canonical orbitals with same spin configuration(s).

eeIdiffSpin::Matrix{T}: Coulomb interactions between canonical orbitals with different spins.

basis::GTBasis{T, D, BN}: The basis set used for the Hartree-Fock approximation.

≡≡≡ Initialization Method(s) ≡≡≡

MatterByHF(HFres::HFfinalVars{T, D, <:Any, NN, BN, HFTS}; 
           roundAtol::Real=getAtolVal(T)) where {T, D, NN, BN, HFTS} -> 
MatterByHF{T, D, NN, <:Any, BN, HFTS}

Construct a MatterByHF from the result of a Hartree-Fock method HFres. Each parameter stored in the constructed CanOrbitals in .occuOrbital and .unocOrbital will be rounded to the nearest multiple of roundAtol; when roundAtol is set to NaN, no rounding will be performed.