Core Types

Quiqbox.ParamBox — Type

ParamBox{T, V, FL} <: DifferentiableParameter{ParamBox, T}

Parameter container that can enable parameter differentiations.

≡≡≡ Field(s) ≡≡≡

data::Array{T, 0}: The data (parameter) stored in a 0-D Array that can be accessed by syntax [].

dataName::Symbol: The name assigned to the stored data.

map::Function: The mathematical mapping of the data. The mapped result can be accessed by syntax ().

canDiff::Array{Bool, 0}: Indicator that whether this container (mapped variable) is marked as "differentiable".

index::Array{<:Union{Int, Nothing}, 0}: Additional index assigned to the parameter.

≡≡≡ Initialization Method(s) ≡≡≡

ParamBox(data::Union{Array{T, 0}, T}, dataName::Symbol=:undef; 
         index::Union{Int, Nothing}=nothing) where {T<:Number} -> 
ParamBox{T, dataName, Quiqbox.FLevel{1, 0}}

ParamBox(data::Union{Array{T, 0}, T}, name::Symbol, mapFunction::Function, 
         dataName::Symbol=:undef; index::Union{Int, Nothing}=nothing, 
         canDiff::Bool=true) where {T<:Number} ->
ParamBox{T, name, Quiqbox.FLevel(mapFunction)}

name specifies the name of the (mapped) variable the ParamBox represents, which helps with symbolic representation and automatic differentiation.

mapFunction: The (mathematical) mapping of the data, which will be stored in the field map. It is for the case where the variable represented by the ParamBox is dependent on another independent variable of which the value is the stored data in the container. After initializing a ParamBox, e.g pb1 = ParamBox(x, mapFunction=f), pb[] returns x, and pb() returns f(x). mapFunction is set to itself in default, which is a dummy function that maps the data to itself.

canDiff determines whether the mapped math variable is "marked" as differentiable (i.e., the mapping is a differentiable function) with respect to the stored data. In other words, it determines whether the mapped variable ParamBox generated during the automatic differentiation procedure is treated as a dependent variable or an independent variable regardless of the mapping relation.

≡≡≡ Example(s) ≡≡≡

julia> ParamBox(1.0)
ParamBox{Float64, :undef, Quiqbox.FLevel{1, 0}}(1.0)[∂][undef]

julia> ParamBox(1.0, :a)
ParamBox{Float64, :a, Quiqbox.FLevel{1, 0}}(1.0)[∂][a]

julia> ParamBox(1.0, :a, abs)
ParamBox{Float64, :a, Quiqbox.FLevel{2, 0}}(1.0)[∂][x_a]

NOTE 1: The rightmost "[∂][IV]" in the printing info indicates the differentiability and the name of the represented independent variable :IV. When the ParamBox is marked as a "differentiable function", "[∂]" is in color green (otherwise it's in grey).

NOTE 2: It's always the (mapped) variable V generated by a ParamBox{<:Any, V} that is used to construct a basis, whereas the underlying independent variable is used to differentiate the basis (in other words, only when mapFunction = itself or canDiff = false is the independent variable same as the mapped variable/represented parameter).

Quiqbox.GaussFunc — Type

GaussFunc <: AbstractGaussFunc

A single contracted gaussian function struct from package Quiqbox.

≡≡≡ Field(s) ≡≡≡

xpn::ParamBox{Float64, :α}：Exponent of the gaussian function.

con::ParamBox{Float64, :d}: Contraction coefficient of the gaussian function.

param::NTuple{2, ParamBox}: A Tuple that stores the ParamBoxs of xpn and con.

≡≡≡ Initialization Method(s) ≡≡≡

GaussFunc(xpn::ParamBox, con::ParamBox) -> GaussFunc

GaussFunc(xpn::Real, con::Real) -> GaussFunc

Quiqbox.BasisFunc — Type

BasisFunc{𝑙, GN} <: FloatingGTBasisFuncs{𝑙, GN, 1}

A (floating) basis function with the center attached to it instead of any nucleus.

≡≡≡ Field(s) ≡≡≡

center::NTuple{3, ParamBox}: The center coordinate in form of a 3-element ParamBox-type Tuple.

gauss::NTuple{N, GaussFunc}: Gaussian functions within the basis function.

ijk::Tuple{XYZTuple{𝑙}}: Cartesian representation (pseudo-quantum number) of the angular momentum orientation. E.g., s (X⁰Y⁰Z⁰) would be XYZTuple{0}((0, 0, 0)). For convenient syntax, .ijk[] converts it to a NTuple{3, Int}.

normalizeGTO::Bool: Whether the GTO::GaussFunc will be normalized in calculations.

param::Tuple{Vararg{<:ParamBox}}： All the tunable parameters::ParamBox stored in the BasisFunc.

≡≡≡ Initialization Method(s) ≡≡≡

BasisFunc(center::Tuple{Vararg{<:ParamBox}}, gauss::NTuple{GN, GaussFunc}, 
          ijk::NTuple{3, Int}, normalizeGTO::Bool) where {GN} -> 
BasisFunc{𝑙, GN} where {𝑙}

BasisFunc(cen::Tuple{Vararg{<:ParamBox}}, gauss::GaussFunc, ijk::NTuple{3, Int}, 
          normalizeGTO::Bool) ->
BasisFunc{𝑙, 1} where {𝑙}

Quiqbox.BasisFuncs — Type

BasisFuncs{𝑙, GN, ON} <: FloatingGTBasisFuncs{𝑙, GN, ON}

A group of basis functions with identical parameters except they have different orientations in the specified subshell. It has the same fields as BasisFunc and specifically, for ijk, the size of the it (ON) can be no less than 1 (and no larger than the size of the corresponding subshell).

Quiqbox.BasisFuncMix — Type

BasisFuncMix{BN, GN} <: CompositeGTBasisFuncs{BN, 1}

Sum of multiple FloatingGTBasisFuncs without any reformulation, treated as one basis Function in the integral calculation.

≡≡≡ Field(s) ≡≡≡

BasisFunc::NTuple{BN, FloatingGTBasisFuncs{<:Any, <:Any, 1}}: Inside basis functions

param::Tuple{Vararg{<:ParamBox}}: Inside parameters.

≡≡≡ Initialization Method(s) ≡≡≡

BasisFuncMix(bfs::Array{<:FloatingGTBasisFuncs{𝑙, GN, 1} where {𝑙, GN}, 1}) ->
BasisFuncMix{BN, GN}

Quiqbox.GTBasis — Type

GTBasis{BT, T} <: BasisSetData{BT}

The container to store basis set information.

≡≡≡ Field(s) ≡≡≡

basis::Vector{<:CompositeGTBasisFuncs{<:Any, 1}}: Basis set.

S::Array{<:Number, 2}: Overlap matrix.

Te::Array{<:Number, 2}: Kinetic energy part of the electronic core Hamiltonian.

eeI::Array{<:Number, 4}: Electron-electron interaction.

getVne::Function: A Function that returns the nuclear attraction Hamiltonian when nuclei::Array{String, 1} and their coordinates::Array{<:AbstractArray, 1} are input.

getHcore::Function: Similar as getVne, a Function that returns the core Hamiltonian when nuclei and their coordinates of same DataType are input.

≡≡≡ Initialization Method(s) ≡≡≡

GTBasis(basis::Array{<:AbstractGTBasisFuncs, 1}, S::Matrix{<:Number}, 
        Te::Matrix{<:Number}, eeI::Array{<:Number, 4}) -> 
GTBasis

GTBasis(basis::Array{<:AbstractGTBasisFuncs, 1}) -> GTBasis

Directly construct a GTBasis given a basis set.

Quiqbox.GridBox — Type

GridBox{NX, NY, NZ} <: SemiMutableParameter{GridBox, Float64}

A struct that stores coordinates of grid points in terms of both Vectors and ParamBoxs.

≡≡≡ Field(s) ≡≡≡

num::Int: Total number of the grid points.

spacing::Float64: The length between adjacent grid points.

box::Vector{NTuple{3, ParamBox}}: The coordinates of grid points in terms of ParamBoxs.

coord::Array{Array{Float64, 1}, 1}: The coordinates of grid points in terms of Vectors.

≡≡≡ Initialization Method(s) ≡≡≡

GridBox(nGrids::NTuple{3, Int}, spacing::Real=10, 
        centerCoord::Array{<:Real, 1}=[0.0,0.0,0.0];
        canDiff::Bool=true, index::Int=0) -> GridBox

Construct a general GridBox that doesn't have to shape as a cube. nGrid is a 3-element Tuple that specifies the number of grids (number of grid points - 1) along 3 dimensions. spacing specifies the length between adjacent grid points. centerCoord specifies the geometry center coordinate of the box. canDiff determines whether the ParamBox should be marked as differentiable. index defines the index number for the actual parameter: spacing L, with the default value 0 it would be L₀.

GridBox(nGridPerEdge::Int, spacing::Real=10, 
        centerCoord::Array{<:Real, 1}=[0.0,0.0,0.0]; 
        canDiff::Bool=true, index::Int=0) -> GridBox

Method of generating a cubic GridBox. nGridPerEdge specifies the number of grids (number of grid points - 1) along each dimension.spacing specifies the length between adjacent grid points. centerCoord specifies the geometry center coordinate of the box. canDiff determines whether the ParamBox should be marked as differentiable. index defines the index number for the actual parameter: spacing L, with the default value 0 it would be L₀.

Quiqbox.HFconfig — Type

HFconfig{HFT, CT} <: ConfigBox{HFconfig, HFT}

The mutable container of Hartree-Fock method configuration.

≡≡≡ Field(s) ≡≡≡

HF::Val{HFT}: Hartree-Fock method type. Available values of HFT are :RHF, :UHF.

C0::Union{Matrix{Float64}, NTuple{2, Matrix{Float64}}, Val}: Initial guess of the coefficient matrix(s) C of the molecular orbitals. When C0 is a Val{T}, the available values of T are :GWH, :Hcore, :SAD.

SCF::SCFconfig: SCF iteration configuration. For more information please refer to SCFconfig.

earlyTermination::Bool: Whether automatically early terminate (skip) a convergence method when its performance becomes unstable or poor.

maxStep::Int: Maximum allowed iteration steps regardless of whether the SCF converges.

≡≡≡ Initialization Method(s) ≡≡≡

HFconfig() -> HFconfig

HFconfig(t::NamedTuple) -> HFconfig

≡≡≡ Example(s) ≡≡≡

julia> HFconfig()
HFconfig{:RHF, :SAD, 3}(Val{:RHF}(), Val{:SAD}(), SCFconfig{3}(interval=(0.0001, 1.0e-6, 1.0e-15), oscillateThreshold=1.0e-5, method, methodConfig)[:ADIIS, :DIIS, :ADIIS], 1000, true)

julia> HFconfig((HF=:UHF,))
HFconfig{:UHF, :SAD, 3}(Val{:UHF}(), Val{:SAD}(), SCFconfig{3}(interval=(0.0001, 1.0e-6, 1.0e-15), oscillateThreshold=1.0e-5, method, methodConfig)[:ADIIS, :DIIS, :ADIIS], 1000, true)

Quiqbox.SCFconfig — Type

SCFconfig{N} <: ImmutableParameter{SCFconfig, Any}

The struct for SCF iteration configurations.

≡≡≡ Field(s) ≡≡≡

method::NTuple{N, Quiqbox.FunctionType}: The applied methods. The available methods and their configurations (in terms of keyword arguments):

Methods	Configuration(s)	keyword argument(s)	Default value(s)

:DS	Damping strength: [0,1]	dampingStrength::Float64	0.0

:DIIS, :EDIIS, :ADIIS	Subspace size (>1); Coefficient solver(:ADMM-> ADMM solver, :LCM -> Lagrange solver)	DIISsize::Int; solver::Symbol	15; :ADMM

interval::NTuple{N, Float64}: The stopping (skipping) thresholds for required methods.

methodConfig::NTuple{N, Vector{<:Pair}}: The additional keywords arguments for each method stored as Tuples of Pairs.

oscillateThreshold::Float64: The threshold for oscillating convergence.

≡≡≡ Initialization Method(s) ≡≡≡

SCFconfig(methods::NTuple{N, Symbol}, intervals::NTuple{N, Float64}, 
          configs::Dict{Int, <:Vector{<:Pair}}=Dict(1=>Pair[]);
          oscillateThreshold::Float64=1e-5) where {N} -> 
SCFconfig{N}

methods and intervals are the methods to be applied and their stopping (skipping) thresholds respectively; the length of those two Vectors should be the same. configs specifies the additional keyword arguments for each methods by a Pair of which the Int key i is for ith method and the pointed Vector{<:Pair} is the pairs of keyword arguments and their values respectively.

≡≡≡ Example(s) ≡≡≡

julia> SCFconfig((:SD, :ADIIS, :DIIS), (1e-4, 1e-12, 1e-13), Dict(2=>[:solver=>:LCM]))
SCFconfig{3}(interval=(0.0001, 1.0e-12, 1.0e-13), oscillateThreshold=1.0e-5, method, methodConfig)[:SD, :ADIIS, :DIIS]

Quiqbox.HFtempVars — Type

HFtempVars{HFT} <: HartreeFockintermediateData

The container to store the intermediate values (only of the same spin configuration) for each iteration during the Hartree-Fock SCF procedure.

≡≡≡ Field(s) ≡≡≡

N::Int: The number of electrons with the same spin function.

Cs::Array{Array{Float64, 2}, 1}: Coefficient matrices.

Fs::Array{Array{Float64, 2}, 1}: Fock matrices

Ds::Array{Array{Float64, 2}, 1}: Density matrices corresponding to only spin configuration. For RHF each elements means (unconverged) 0.5*Dᵀ.

Es::Array{Float64, 1}: Part of Hartree-Fock energy corresponding to only spin configuration. For RHF each element means (unconverged) 0.5*E0HF.

shared.Dtots::Array{Array{Float64, 2}, 1}: The total density matrices.

shared.Etots::Array{Float64, 1}: The total Hartree-Fock energy.

NOTE: For UHF, there are 2 HFtempVars being updated during the SCF iterations, and change the field shared.Dtots or shared.Etots of one container will affect the other one's.

Quiqbox.HFfinalVars — Type

HFfinalVars{HFT} <: HartreeFockFinalValue{HFT}

The container of the final values after a Hartree-Fock SCF procedure.

≡≡≡ Field(s) ≡≡≡

E0HF::Float64: Hartree-Fock energy of the electronic Hamiltonian.

N::Int: The total number of electrons.

C::Union{Array{Float64, 2}, NTuple{2, Array{Float64, 2}}}: Coefficient matrix(s) for one spin configuration.

F::Union{Array{Float64, 2}, NTuple{2, Array{Float64, 2}}}: Fock matrix(s) for one spin configuration.

D::Union{Array{Float64, 2}, NTuple{2, Array{Float64, 2}}}: Density matrix(s) for one spin configuration.

Emo::Union{Array{Float64, 1}, NTuple{2, Array{Float64, 1}}}: Energies of molecular orbitals.

occu::Union{Array{Int, 1}, NTuple{2, Array{Int, 1}}}: occupation numbers of molecular orbitals.

temp::Union{HFtempVars{HFT}, NTuple{2, HFtempVars{HFT}}} the intermediate values.

isConverged::Bool: Whether the SCF procedure is converged in the end.

Quiqbox.Molecule — Type

Molecule{Nc, Ne, Nb} <:MolecularHartreeFockCoefficient{Nc, Ne}

Container for the information of a molecule.

≡≡≡ Field(s) ≡≡≡

nuc::Tuple{Vararg{String}}: Nuclei of the molecule.

nucCoords::Tuple{Vararg{NTuple{3, Real}}}: The coordinates of the nuclei.

Ne::Int: The number of electrons.

orbital::Tuple{Vararg{MolOrbital}}: Molecular orbitals.

basis::Tuple{Vararg{FloatingGTBasisFuncs}}: The basis set for the molecular orbitals.

E0HF::Float64: Hartree-Fock energy of the electronic Hamiltonian from the basis set.

EnnR::Float64: The nuclear-nuclear repulsion energy.

≡≡≡ Initialization Method(s) ≡≡≡

Molecule(basis::Array{FloatingGTBasisFuncs, 1}, nuc::Array{String, 1}, 
         nucCoords::Array{<:AbstractArray, 1}, Ne::Int, E0HF::Float64, 
         Emos::Array{Float64, 1}, occus::Array{<:Real, 1}, C::Array{Float64, 2}, 
         spins::Array{String, 1}, 
         symms::Array{String, 1}=repeat(["A"], length(occus))) -> 
Molecule{<:Any, Ne, <:Any}

Emos are the energies of corresponding molecular energies. occus are the occupation numbers of the orbitals. C is the coefficient matrix, which does not need to be a square matrix since the number of rows is the size of the (spatial) basis set whereas the number of the columns represents the number of molecular orbitals. spin specifies the spin functions of the orbitals, entries of which can be set to "Alpha" or "Beta". symms are symmetries of the orbitals where the default entry value is "A" for being antisymmetric.

Molecule(basis::Array{<:FloatingGTBasisFuncs, 1}, nuc::Array{String, 1}, 
         nucCoords::Array{<:AbstractArray, 1}, HFfVars::HFfinalVars) -> 
Molecule

Construct a Molecule from a basis set, nuclei information, and the result from the corresponding Hartree-Fock SCF procedure, specifically a HFfinalVars struct.

Quiqbox.MolOrbital — Type

MolOrbital{N} <: AbstractMolOrbital

Struct of molecular orbitals.

≡≡≡ Field(s) ≡≡≡

symmetry::String: The symmetry of the orbital. The default value is "A" for being antisymmetric.

energy::Float64: Molecular energy.

spin::String: Spin function of the orbital. Available values: "Alpha", "Beta".

occupancy::Real: Occupation number.

orbitalCoeffs::NTuple{N, Float64}: coefficients of the basis functions to form the molecular orbital.

≡≡≡ Initialization Method(s) ≡≡≡

MolOrbital(energy::Float64, occupancy::Real, orbitalCoeffs::Array{Float64, 1}, 
           spin::String="Alpha", symmetry::String="A") -> MolOrbital{N}

Quiqbox.POconfig — Type

POconfig{M, T, F} <: ConfigBox{POconfig, M}

≡≡≡ Field(s) ≡≡≡

The mutable container of parameter optimization configuration.

method::Val{M}: The method to calculate objective function (e.g., HF energy) for optimization. Available values of M from Quiqbox are :HF,.

config::ConfigBox{<:Any, T}: The configuration struct for the selected method. E.g., for :HF it's HFconfig.

target::Float64: The value of target function aimed to achieve.

error::Float64: The error for the convergence when evaluating difference between the latest few energies. When set to NaN, there will be no convergence detection.

maxStep::Int: Maximum allowed iteration steps regardless of whether the optimization iteration converges.

GD::F1: Applied gradient descent Function. Default method is gradDescent!.

≡≡≡ Initialization Method(s) ≡≡≡

POconfig() -> POconfig

POconfig(t::NamedTuple) -> POconfig

≡≡≡ Example(s) ≡≡≡

julia> POconfig()
POconfig{:HF, HFconfig{:RHF, :SAD, 3}, typeof(gradDescent!)}(Val{:HF}(), HFconfig{:RHF, :SAD, 3}(Val{:RHF}(), Val{:SAD}(), SCFconfig{3}(interval=(0.0001, 1.0e-6, 1.0e-15), oscillateThreshold=1.0e-5, method, methodConfig)[:ADIIS, :DIIS, :ADIIS], 1000, true), NaN, 1.0e-5, 500, Quiqbox.gradDescent!)

julia> POconfig((maxStep=100,))
POconfig{:HF, HFconfig{:RHF, :SAD, 3}, typeof(gradDescent!)}(Val{:HF}(), HFconfig{:RHF, :SAD, 3}(Val{:RHF}(), Val{:SAD}(), SCFconfig{3}(interval=(0.0001, 1.0e-6, 1.0e-15), oscillateThreshold=1.0e-5, method, methodConfig)[:ADIIS, :DIIS, :ADIIS], 1000, true), NaN, 1.0e-5, 100, Quiqbox.gradDescent!)